splitting up codegen
