+++ src/liblesma/Backend/CodegenVisit.cpp
@@ -70,6 +70,7 @@ LLD_HAS_DRIVER(elf)
 
 #include "liblesma/AST/AST.h"
 #include "liblesma/Backend/CodegenError.h"
+#include "liblesma/Backend/CodegenRuntimeNames.h"
 #include "liblesma/Backend/CodegenTypeUtils.h"
 #include "liblesma/Backend/MangleUtils.h"
 #include "liblesma/Common/ExportDiscovery.h"
@@ -135,84 +136,6 @@ Codegen::Codegen(std::shared_ptr<Parser> parser, std::shared_ptr<SourceMgr> srcM
   topLevelFunc = initializeTopLevel();
   // base.les is loaded at the start of run() so we don't load it during
   // constructor re-entrancy when creating Codegens for imported modules.
-}
-
-auto Codegen::initializeModule() -> std::unique_ptr<Module> {
-  std::unique_ptr<Module> mod;
-#if LLVM_VERSION_MAJOR >= 21
-  theContext->withContextDo(
-      [&mod](LLVMContext* ctx) { mod = std::make_unique<Module>("Lesma", *ctx); });
-#else
-  {
-    auto lock = theContext->getLock();
-    LLVMContext* ctx = theContext->getContext();
-    mod = std::make_unique<Module>("Lesma", *ctx);
-  }
-#endif
-#if LLVM_VERSION_MAJOR >= 21
-  mod->setTargetTriple(targetMachine->getTargetTriple());
-#else
-  mod->setTargetTriple(targetMachine->getTargetTriple().str());
-#endif
-  mod->setDataLayout(targetMachine->createDataLayout());
-  mod->setSourceFileName(filename);
-
-  return mod;
-}
-
-// NOLINTNEXTLINE(readability-convert-member-functions-to-static)
-auto Codegen::initializeTargetMachine() -> std::unique_ptr<llvm::TargetMachine> {
-  // Configure output target
-  auto targetTriple = llvm::Triple(llvm::sys::getDefaultTargetTriple());
-
-  // Search after selected target
-  std::string error;
-  const llvm::Target* target = llvm::TargetRegistry::lookupTarget(targetTriple.getTriple(), error);
-  if (target == nullptr) {
-    throw CodegenError({}, "Target not available:\n{}", error);
-  }
-
-  llvm::TargetOptions const opt;
-  llvm::Reloc::Model rm = llvm::Reloc::Model();
-  std::unique_ptr<llvm::TargetMachine> targetMachine(
-#if LLVM_VERSION_MAJOR >= 21
-      target->createTargetMachine(targetTriple, "generic", "", opt, rm));
-#else
-      target->createTargetMachine(targetTriple.str(), "generic", "", opt, rm));
-#endif
-  return targetMachine;
-}
-
-auto Codegen::initializeJit() -> std::unique_ptr<LLJIT> {
-  llvm::orc::LLJITBuilder jitBuilder{};
-  jitBuilder.setDataLayout(theModule->getDataLayout());
-  jitBuilder.setJITTargetMachineBuilder(
-      llvm::orc::JITTargetMachineBuilder(targetMachine->getTargetTriple()));
-  auto jit = llvm::cantFail(jitBuilder.create());
-  if (!jit) {
-    throw CodegenError({}, "Couldn't initialize JIT\n");
-  }
-
-  // Add support for C native functions
-  auto& mainJd = jit->getMainJITDylib();
-  auto generator = cantFail(
-      DynamicLibrarySearchGenerator::GetForCurrentProcess(jit->getDataLayout().getGlobalPrefix()));
-  mainJd.addGenerator(std::move(generator));
-
-  return jit;
-}
-
-auto Codegen::initializeTopLevel() -> llvm::Function* {
-  std::vector<llvm::Type*> const paramTypes = {};
-
-  FunctionType* ft = FunctionType::get(builder->getInt64Ty(), paramTypes, false);
-  Function* f = Function::Create(ft, isMain ? Function::ExternalLinkage : Function::InternalLinkage,
-                                 "main", *theModule);
-
-  auto* entry = BasicBlock::Create(theModule->getContext(), "entry", f);
-  builder->SetInsertPoint(entry);
-
-  return f;
 }
 
 auto Codegen::defineFunction(lesma::Value* value, const FuncDecl* node, Value* clsSymbol) -> void {
@@ -334,485 +257,6 @@ auto Codegen::defineFunction(lesma::Value* value, const FuncDecl* node, Value* c
   builder->SetInsertPoint(&topLevelFunc->back());
 }
 
-auto Codegen::getExportsFromFile(const std::string& filepath, bool isStd,
-                                 const std::string& mainFilePath) -> std::vector<std::string> {
-  return getExportedTopLevelNamesFromFile(filepath, isStd, mainFilePath);
-}
-
-auto Codegen::typecheckModule(const Compound* ast, const std::string& modulePath)
-    -> std::tuple<std::unique_ptr<SymbolTable>, std::vector<std::unique_ptr<lesma::Type>>,
-                  std::unordered_map<lesma::Type*, std::unordered_map<std::string, lesma::Type*>>> {
-  Typechecker typechecker(
-      modulePath, [this](const std::string& path, bool isStd, const std::string& mainFilePath) {
-        return getExportsFromFile(path, isStd, mainFilePath);
-      });
-  typechecker.run(ast);
-  return {typechecker.takeRootScope(), typechecker.takeTypeCache(),
-          typechecker.takeSpecializedTypeEnv()};
-}
-
-auto Codegen::isImported(const std::vector<ImportedNameBinding>& importedNames,
-                         const std::string& importName) const -> bool {
-  return std::ranges::any_of(importedNames, [&importName](const ImportedNameBinding& binding) {
-    return binding.name == importName;
-  });
-}
-
-auto Codegen::getImportedLocalName(const std::vector<ImportedNameBinding>& importedNames,
-                                   const std::string& importName) const -> std::string {
-  for (const ImportedNameBinding& binding : importedNames) {
-    if (binding.name == importName) {
-      return binding.alias.empty() ? importName : binding.alias;
-    }
-  }
-  return "";
-}
-
-auto Codegen::insertImportAlias(const std::string& moduleAlias, bool importToScope) -> void {
-  if (importToScope) {
-    return;
-  }
-
-  auto importTyp = std::make_unique<Type>(BaseType::TY_IMPORT);
-  auto* importTypPtr = importTyp.get();
-  auto importSym = std::make_unique<Value>(moduleAlias, importTypPtr);
-  importSym->setCategory(ValueCategory::MODULE_SYMBOL);
-  scope->insertSymbol(std::move(importSym));
-  scope->insertType(moduleAlias, std::move(importTyp));
-}
-
-auto Codegen::exposeImportedSymbols(llvm::SMRange /*span*/, SymbolTable* importedScope,
-                                    bool importAll, bool importToScope,
-                                    const std::vector<ImportedNameBinding>& importedNames) -> void {
-  for (auto* sym : importedScope->getSymbols()) {
-    const bool importedByName = isImported(importedNames, sym->getName());
-    const std::string importedLocalName = getImportedLocalName(importedNames, sym->getName());
-    const bool exposeClassForModuleImport =
-        !importToScope && sym->getType()->is(BaseType::TY_CLASS);
-    // Class/enum types from the module are always merged for lookup: exported APIs
-    // may reference non-exported helper types in signatures.
-    if (sym->getType()->isOneOf({BaseType::TY_ENUM, BaseType::TY_CLASS}) &&
-        (importAll || importedByName || exposeClassForModuleImport)) {
-      llvm::StructType* structType =
-          StructType::getTypeByName(theModule->getContext(), sym->getName());
-      const std::string localName = importedLocalName.empty() ? sym->getName() : importedLocalName;
-      auto structSymbol = std::make_unique<Value>(localName, sym->getType());
-      structSymbol->setCategory(ValueCategory::TYPE_SYMBOL);
-      structSymbol->getType()->setLlvmType(structType);
-      structSymbol->setGenericClassTemplate(sym->getGenericClassTemplate());
-      scope->insertTypeRef(sym->getName(), sym->getType());
-      scope->insertSymbol(std::move(structSymbol));
-      continue;
-    }
-
-    if (!sym->getType()->is(BaseType::TY_FUNCTION) || !sym->isExported()) {
-      continue;
-    }
-
-    if (sym->getType()->getLlvmType() == nullptr) {
-      continue;
-    }
-
-    auto* fTy = llvm::cast<FunctionType>(sym->getType()->getLlvmType());
-    auto name = sym->getName();
-    std::vector<lesma::Type*> paramTypes;
-    for (auto* field : sym->getType()->getFields()) {
-      paramTypes.push_back(field->type);
-    }
-
-    Value* funcSymbol = importedScope->lookupFunction(name, paramTypes);
-    const bool isMethodSym = MangleUtils::isMethod(sym->getMangledName());
-    bool methodClassImported = true;
-    if (isMethodSym && !importAll && importToScope) {
-      const std::string& mn = sym->getMangledName();
-      const size_t colcol = mn.find("::");
-      if (colcol != std::string::npos) {
-        std::string classPart = mn.substr(0, colcol);
-        const size_t arrow = classPart.find("=>");
-        if (arrow != std::string::npos) {
-          classPart = classPart.substr(arrow + 2);
-        }
-        methodClassImported = isImported(importedNames, classPart);
-      }
-    }
-    if (funcSymbol == nullptr || !funcSymbol->isExported() ||
-        (!importAll && !importedByName &&
-         (!isMethodSym || (importToScope && !methodClassImported)))) {
-      continue;
-    }
-
-    const std::string localName = importedLocalName.empty() ? name : importedLocalName;
-    Value* localSymbol = scope->lookupFunction(localName, paramTypes);
-    const bool reuseExistingLocal =
-        localSymbol != nullptr && localSymbol->getLlvmValue() == nullptr;
-    auto symbol =
-        reuseExistingLocal ? nullptr : std::make_unique<Value>(localName, funcSymbol->getType());
-    Value* targetSymbol = reuseExistingLocal ? localSymbol : symbol.get();
-    targetSymbol->setCategory(ValueCategory::CALLABLE_SYMBOL);
-    if (isJit) {
-      auto* f = llvm::cast<Function>(
-          theModule->getOrInsertFunction(sym->getMangledName(), fTy).getCallee());
-      targetSymbol->setType(funcSymbol->getType());
-      targetSymbol->getType()->setLlvmType(fTy);
-      targetSymbol->setLlvmValue(f);
-    } else {
-      auto* newFunc =
-          Function::Create(fTy, Function::ExternalLinkage, sym->getMangledName(), *theModule);
-      targetSymbol->setType(funcSymbol->getType());
-      targetSymbol->getType()->setLlvmType(newFunc->getFunctionType());
-      targetSymbol->setLlvmValue(newFunc);
-    }
-    targetSymbol->setExported(false);
-    targetSymbol->setMangledName(sym->getMangledName());
-    if (!reuseExistingLocal) {
-      scope->insertSymbol(std::move(symbol));
-    }
-  }
-}
-
-auto Codegen::compileModule(llvm::SMRange span, const std::string& filepath, bool isStd,
-                            const std::string& moduleAlias, bool importAll, bool importToScope,
-                            const std::vector<ImportedNameBinding>& importedNames) -> void {
-  std::filesystem::path mainPath = filename;
-  // Read source
-  auto absolutePath =
-      isStd ? filepath
-            : fmt::format("{}/{}", std::filesystem::absolute(mainPath).parent_path().c_str(),
-                          filepath);
-
-  static thread_local std::unordered_set<std::string> compiling;
-  if (compiling.contains(absolutePath)) {
-    throw CodegenError(span, "Circular import detected: {}", filepath);
-  }
-
-  // If this codegen already compiled this module, only merge its symbols (no
-  // recompilation). Only skip when we have the scope (same codegen compiled
-  // it); otherwise a child may see the path in the list but not have the scope.
-  auto it = std::find(importedModules->begin(), importedModules->end(), absolutePath);
-  if (it != importedModules->end()) {
-    auto existingIdx = static_cast<size_t>(it - importedModules->begin());
-    if (existingIdx >= importedScopes->size()) {
-      throw CodegenError(span, "Circular import detected: {}", filepath);
-    }
-  }
-  if (it != importedModules->end()) {
-    auto existingIdx = static_cast<size_t>(it - importedModules->begin());
-    SymbolTable* existingScope = importedScopes->at(existingIdx).get();
-    if (existingIdx < importedCodegens.size() && importedCodegens[existingIdx] != nullptr) {
-      mergeImportedTraitMetadata(*importedCodegens[existingIdx]);
-      mergeImportedSpecializationState(*importedCodegens[existingIdx]);
-    }
-    insertImportAlias(moduleAlias, importToScope);
-    exposeImportedSymbols(span, existingScope, importAll, importToScope, importedNames);
-    return;
-  }
-
-  // Guard against circular import: if this module is already being compiled
-  // (path in importedModules but scope not yet in importedScopes), we would
-  // fall through and re-compile it → infinite recursion. Detect and error.
-  if (compiling.size() >= 32) {
-    throw CodegenError(span,
-                       "Import recursion depth exceeded (possible circular "
-                       "import): {}",
-                       filepath);
-  }
-  compiling.insert(absolutePath);
-
-  auto buffer = MemoryBuffer::getFile(absolutePath);
-  if (std::error_code ec = buffer.getError()) {
-    compiling.erase(absolutePath);
-    throw LesmaError(llvm::SMRange(), "Could not read file: {}", absolutePath);
-  }
-
-  auto fileId = sourceManager->AddNewSourceBuffer(std::move(*buffer), llvm::SMLoc());
-  // auto sourceStr =
-  // sourceManager->getMemoryBuffer(fileId)->getBuffer().str();
-  importedModules->push_back(absolutePath);
-
-  try {
-    // Lexer
-    auto lexer = std::make_unique<Lexer>(sourceManager);
-    lexer->scanAll();
-
-    // Parser
-    auto parser = std::make_unique<Parser>(lexer->getTokens());
-    parser->parse();
-    Compound* ast = parser->getAst();
-    if (ast == nullptr) {
-      throw CodegenError(span, "Unable to parse imported module {}", filepath);
-    }
-
-    auto [preScope, preTypeCache, preSpecEnv] = typecheckModule(ast, absolutePath);
-
-    auto codegen = std::make_unique<Codegen>(
-        std::move(parser), sourceManager, absolutePath, std::vector<std::string>{}, isJit, false,
-        !importToScope ? moduleAlias : "", theContext, importedModules, importedScopes,
-        std::move(preScope), std::move(preTypeCache), std::move(preSpecEnv));
-    codegen->run();
-    mergeImportedTraitMetadata(*codegen);
-
-    // O0: imported modules are separate LLVM modules; O3+ADCE can drop defs only referenced
-    // from another module (e.g. list methods). The main module is optimized in Driver.
-    codegen->optimize(OptimizationLevel::O0);
-    codegen->theModule->setModuleIdentifier(filepath);
-
-    insertImportAlias(moduleAlias, importToScope);
-    exposeImportedSymbols(span, codegen->rootScope.get(), importAll, importToScope, importedNames);
-
-    importedScopes->push_back(std::move(codegen->rootScope));
-    codegen->scope = nullptr; // Clear navigation pointer (rootscope now moved)
-
-    for (auto& type : codegen->typeCache) {
-      typeCache.push_back(std::move(type));
-    }
-    mergeImportedSpecializationState(*codegen);
-
-    if (isJit) {
-      // Add the module to JIT (after importing symbols; theModule still valid)
-      cantFail(theJit->addIRModule(ThreadSafeModule(std::move(codegen->theModule), *theContext)),
-               fmt::format("Failed adding import {} to JIT", filename).c_str());
-    } else {
-      // Create object file to be linked
-      std::string objFile = fmt::format("tmp{}", objectFiles.size());
-      codegen->writeToObjectFile(objFile);
-      objectFiles.push_back(fmt::format("{}.o", objFile));
-    }
-    // Keep the imported codegen alive so Class* stored in copied symbols stay
-    // valid
-    importedCodegens.push_back(std::move(codegen));
-  } catch (const LesmaError& err) {
-    compiling.erase(absolutePath);
-    if (!err.getSpan().isValid()) {
-      lesma::print(LogType::ERROR, err.what());
-    } else {
-      showInline(sourceManager.get(), fileId, err.getSpan(), absolutePath, true, err.what());
-    }
-
-    throw CodegenError(span, "Unable to import {} due to errors", filepath);
-  }
-  compiling.erase(absolutePath);
-}
-
-auto Codegen::optimize(OptimizationLevel opt) -> void {
-  if (opt == OptimizationLevel::O0) {
-    return;
-  }
-
-  llvm::LoopAnalysisManager lam;
-  llvm::FunctionAnalysisManager fam;
-  llvm::CGSCCAnalysisManager cgam;
-  llvm::ModuleAnalysisManager mam;
-
-  llvm::PassBuilder pb(&*targetMachine);
-
-  pb.registerModuleAnalyses(mam);
-  pb.registerCGSCCAnalyses(cgam);
-  pb.registerFunctionAnalyses(fam);
-  pb.registerLoopAnalyses(lam);
-  pb.crossRegisterProxies(lam, fam, cgam, mam);
-
-  // Add custom passes to LoopPassManager
-  llvm::LoopPassManager lpm;
-  lpm.addPass(llvm::LoopFullUnrollPass());
-
-  // Add custom passes to FunctionPassManager
-  llvm::FunctionPassManager fpm;
-  fpm.addPass(llvm::ADCEPass());
-  fpm.addPass(llvm::GVNPass());
-  fpm.addPass(llvm::DSEPass());
-  fpm.addPass(llvm::LoopVectorizePass());
-  fpm.addPass(llvm::createFunctionToLoopPassAdaptor(std::move(lpm)));
-
-  // Add custom passes to CGSCCPassManager
-  llvm::CGSCCPassManager cgpm;
-  cgpm.addPass(llvm::InlinerPass());
-
-  // Add custom pass managers to ModulePassManager
-  llvm::ModulePassManager mpm =
-      pb.buildModuleOptimizationPipeline(opt, ThinOrFullLTOPhase::FullLTOPreLink);
-  mpm.addPass(llvm::createModuleToPostOrderCGSCCPassAdaptor(std::move(cgpm)));
-  mpm.addPass(llvm::createModuleToFunctionPassAdaptor(std::move(fpm)));
-  mpm.addPass(llvm::GlobalDCEPass());
-
-  mpm.run(*theModule, mam);
-}
-
-auto Codegen::writeToObjectFile(const std::string& output) -> void {
-  std::error_code err;
-  llvm::raw_fd_ostream out(output + ".o", err);
-
-  if (err) {
-    throw CodegenError({}, "Error opening file {} for writing: {}", output, err.message());
-  }
-
-  llvm::legacy::PassManager passManager;
-  if (targetMachine->addPassesToEmitFile(passManager, out, nullptr,
-                                         llvm::CodeGenFileType::ObjectFile)) {
-    throw CodegenError({}, "Target Machine can't emit an object file");
-  }
-  // Emit object file
-  passManager.run(*theModule);
-
-  // Flush and close the file
-  out.flush();
-  out.close();
-}
-
-void Codegen::linkObjectFileWithLld(const std::string& objFilename) {
-  std::string output = getBasename(objFilename);
-
-  std::vector<const char*> args;
-
-  // First arg determines linker flavor: ld.lld (ELF), ld64.lld (MachO),
-  // lld-link (COFF)
-#ifdef __APPLE__
-  args.push_back("ld64.lld");
-#elif defined(_WIN32)
-  args.push_back("lld-link");
-#else
-  args.push_back("ld.lld");
-#endif
-
-  // Suppress linker warnings
-  args.push_back("-w");
-  // Files
-  args.push_back("-o");
-  args.push_back(output.c_str());
-  args.push_back(objFilename.c_str());
-  for (const auto& obj : objectFiles) {
-    args.push_back(obj.c_str());
-  }
-
-#ifdef __APPLE__
-  // Add macOS-specific linker arguments
-  args.push_back("-arch");
-  args.push_back("arm64");
-  args.push_back("-platform_version");
-  args.push_back("macos"); // platform
-  args.push_back("11.0");  // min version
-  args.push_back("11.0");  // sdk version
-  args.push_back("-L");
-  args.push_back("/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/lib");
-  args.push_back("-lSystem");
-#endif
-
-  // Run the LLD linker using lldMain
-#ifdef __APPLE__
-  lld::Result result =
-      lld::lldMain(args, llvm::outs(), llvm::errs(), {{.f = lld::Darwin, .d = &lld::macho::link}});
-#elif defined(_WIN32)
-  lld::Result result =
-      lld::lldMain(args, llvm::outs(), llvm::errs(), {{.f = lld::WinLink, .d = &lld::coff::link}});
-#else
-  lld::Result result =
-      lld::lldMain(args, llvm::outs(), llvm::errs(), {{.f = lld::Gnu, .d = &lld::elf::link}});
-#endif
-
-  if (result.retCode != 0) {
-    throw CodegenError({}, "Linking Failed");
-  }
-
-  // Remove object files (ignore errors as they're temporary)
-  std::ignore = llvm::sys::fs::remove(objFilename);
-  for (const auto& obj : objectFiles) {
-    std::ignore = llvm::sys::fs::remove(obj);
-  }
-}
-
-auto Codegen::linkObjectFile(const std::string& objFilename) -> void {
-  linkObjectFileWithLld(objFilename);
-}
-
-auto Codegen::prepareJit() -> void {
-  auto jitError = theJit->addIRModule(ThreadSafeModule(std::move(theModule), *theContext));
-  if (jitError) {
-    throw CodegenError({}, "JIT Error:\n{}");
-  }
-  auto mainFunc = theJit->lookup(topLevelFunc->getName());
-  if (!mainFunc) {
-    throw CodegenError({}, "Couldn't find top level function\n");
-  }
-  mainFuncAddress = mainFunc->toPtr<MainFnTy>();
-}
-
-auto Codegen::executeJit() -> int {
-  if (mainFuncAddress == nullptr) {
-    throw CodegenError({}, "Main function address not found, did you prepare JIT?\n");
-  }
-
-  return mainFuncAddress();
-}
-
-auto Codegen::run() -> void {
-  collectTraitMetadataFromAst();
-  // Load implicit stdlib modules once so every module has base (including list) helpers.
-  // Done here (not in constructor) to avoid re-entrancy when creating Codegens for imported
-  // modules.
-  std::vector<std::string> const implicitStdlibModules = {"base.les"};
-  auto const currentPath =
-      std::filesystem::absolute(std::filesystem::path(filename)).lexically_normal();
-  auto const basePath =
-      std::filesystem::absolute(std::filesystem::path(getStdDir()) / "base.les").lexically_normal();
-  const bool mainIsStdlibEntry = !filename.empty() && currentPath == basePath;
-  if (!mainIsStdlibEntry || filename.empty()) {
-    for (const auto& moduleName : implicitStdlibModules) {
-      auto const modulePath =
-          std::filesystem::absolute(std::filesystem::path(getStdDir()) / moduleName)
-              .lexically_normal();
-      if (currentPath == modulePath) {
-        continue;
-      }
-      compileModule(llvm::SMRange(), getStdDir() + moduleName, true, moduleName, true, true, {});
-    }
-  }
-
-  deferStack.emplace();
-  parser->getAst()->accept(*this);
-
-  auto instrs = deferStack.top();
-  deferStack.pop();
-
-  // Visit all statements
-  for (auto* inst : instrs) {
-    inst->accept(*this);
-  }
-
-  // Define the function bodies (index-based: specializeFunction may append
-  // new prototypes while we iterate, e.g. when combine<int> triggers add<int>)
-  for (size_t pi = 0; pi < prototypes.size(); ++pi) {
-    auto* fn = std::get<0>(prototypes[pi]);
-    auto savedGenerics = currentGenericTypes;
-    if (auto env = specializationEnvs.find(fn); env != specializationEnvs.end()) {
-      currentGenericTypes = env->second;
-    } else if (auto* cls = std::get<2>(prototypes[pi]);
-               cls != nullptr && cls->getType() != nullptr &&
-               cls->getType()->is(BaseType::TY_PTR) &&
-               cls->getType()->getElementType() != nullptr) {
-      if (auto clsEnv = specializedClassTypeEnvs.find(cls->getType()->getElementType());
-          clsEnv != specializedClassTypeEnvs.end()) {
-        currentGenericTypes = clsEnv->second;
-      }
-    } else {
-      auto fields = fn->getType()->getFields();
-      if (!fields.empty() && fields.front()->type != nullptr &&
-          fields.front()->type->is(BaseType::TY_PTR) &&
-          fields.front()->type->getElementType() != nullptr) {
-        if (auto clsEnv = specializedClassTypeEnvs.find(fields.front()->type->getElementType());
-            clsEnv != specializedClassTypeEnvs.end()) {
-          currentGenericTypes = clsEnv->second;
-        }
-      }
-    }
-    defineFunction(fn, std::get<1>(prototypes[pi]), std::get<2>(prototypes[pi]));
-    currentGenericTypes = std::move(savedGenerics);
-  }
-
-  // Return 0 for top-level function
-  builder->CreateRet(ConstantInt::getSigned(builder->getInt64Ty(), 0));
-}
-
-auto Codegen::dump() -> void { theModule->print(outs(), nullptr); }
-
 auto Codegen::visit(const Statement* node) -> void {
   lesma::print("Visited a blank statement\n{}", node->toString(sourceManager.get(), "", true));
 }
@@ -821,1476 +265,6 @@ auto Codegen::visit(const Expression* node) -> void {
   lesma::print("Visited a blank expression\n{}", node->toString(sourceManager.get(), "", true));
 }
 
-auto Codegen::visit(const TypeExpr* node) -> void {
-  // For primitive types, cache them so they survive beyond result's lifetime
-  // This is needed because setReturnType and similar store raw Type* pointers
-  if (node->getType() == TokenType::INT_TYPE) {
-    auto* type = cacheType(std::make_unique<Type>(BaseType::TY_INT, builder->getInt64Ty()));
-    type->setIntWidth(64);
-    result = std::make_unique<Value>(type);
-  } else if (node->getType() == TokenType::INT8_TYPE) {
-    auto* type = cacheType(std::make_unique<Type>(BaseType::TY_INT, builder->getInt8Ty()));
-    type->setIntWidth(8);
-    result = std::make_unique<Value>(type);
-  } else if (node->getType() == TokenType::INT16_TYPE) {
-    auto* type = cacheType(std::make_unique<Type>(BaseType::TY_INT, builder->getInt16Ty()));
-    type->setIntWidth(16);
-    result = std::make_unique<Value>(type);
-  } else if (node->getType() == TokenType::INT32_TYPE) {
-    auto* type = cacheType(std::make_unique<Type>(BaseType::TY_INT, builder->getInt32Ty()));
-    type->setIntWidth(32);
-    result = std::make_unique<Value>(type);
-  } else if (node->getType() == TokenType::FLOAT_TYPE) {
-    auto* type = cacheType(std::make_unique<Type>(BaseType::TY_FLOAT, builder->getDoubleTy()));
-    result = std::make_unique<Value>(type);
-  } else if (node->getType() == TokenType::FLOAT32_TYPE) {
-    auto* type = cacheType(std::make_unique<Type>(BaseType::TY_FLOAT, builder->getFloatTy()));
-    result = std::make_unique<Value>(type);
-  } else if (node->getType() == TokenType::BOOL_TYPE) {
-    auto* type = cacheType(std::make_unique<Type>(BaseType::TY_BOOL, builder->getInt1Ty()));
-    result = std::make_unique<Value>(type);
-  } else if (node->getType() == TokenType::STRING_TYPE) {
-    auto* type = cacheType(std::make_unique<Type>(BaseType::TY_STRING, builder->getPtrTy()));
-    result = std::make_unique<Value>(type);
-  } else if (node->getType() == TokenType::VOID_TYPE) {
-    auto* type = cacheType(std::make_unique<Type>(BaseType::TY_VOID, builder->getVoidTy()));
-    result = std::make_unique<Value>(type);
-  } else if (node->getType() == TokenType::PTR_TYPE) {
-    node->getElementType()->accept(*this);
-    // Function type is already a pointer at LLVM level; parser uses *func for
-    // consistency, so do not add another pointer layer.
-    if (!result->getType()->is(BaseType::TY_FUNCTION)) {
-      auto* type = cacheType(
-          std::make_unique<Type>(BaseType::TY_PTR, builder->getPtrTy(), result->getType()));
-      result = std::make_unique<Value>(type);
-    }
-  } else if (node->getType() == TokenType::FUNC_TYPE) {
-    node->getReturnType()->accept(*this);
-    auto retType = std::move(result);
-    std::vector<std::unique_ptr<Field>> fields;
-    std::vector<lesma::Type*> paramTypes;
-    std::vector<llvm::Type*> paramLLVMTypes;
-    for (auto* paramType : node->getParams()) {
-      paramType->accept(*this);
-      paramLLVMTypes.push_back(result->getType()->getLlvmType());
-      paramTypes.push_back(result->getType());
-      fields.push_back(std::make_unique<Field>(result->getName(), result->getType()));
-    }
-
-    // With opaque pointers, function pointer types are just `ptr`
-    // The actual function signature is tracked in Lesma's Type system via
-    // fields
-    auto funcType =
-        std::make_unique<Type>(BaseType::TY_FUNCTION, builder->getPtrTy(), std::move(fields));
-    funcType->setReturnType(retType->getType());
-    result = std::make_unique<Value>(cacheType(std::move(funcType)));
-  } else if (node->getType() == TokenType::CUSTOM_TYPE) {
-    const std::string lookupName = node->getLookupName();
-    auto git = currentGenericTypes.find(lookupName);
-    if (git != currentGenericTypes.end()) {
-      result = std::make_unique<Value>(git->second);
-      return;
-    }
-    std::vector<lesma::Type*> explicitTypeArgs;
-    for (auto* typeArg : node->getTypeArgs()) {
-      typeArg->accept(*this);
-      explicitTypeArgs.push_back(result->getType());
-    }
-    if (lookupName == "__buffer") {
-      if (explicitTypeArgs.size() != 1U) {
-        throw CodegenError(node->getSpan(), "__buffer<T> expects exactly one type argument");
-      }
-      auto* type =
-          cacheType(std::make_unique<Type>(BaseType::TY_ARRAY, nullptr, explicitTypeArgs.front()));
-      type->setDisplayName(node->getName());
-      getOrCreateLlvmType(type);
-      result = std::make_unique<Value>(type);
-      return;
-    }
-    auto* typ = scope->lookupType(lookupName);
-    auto* sym = scope->lookupStruct(lookupName);
-    if (typ == nullptr && sym == nullptr) {
-      throw CodegenError(node->getSpan(), "Type not found: {}", node->getName());
-    }
-    if (typ != nullptr && typ->is(BaseType::TY_TRAIT_EXISTENTIAL)) {
-      getOrCreateLlvmType(typ);
-      Value* traitSym = scope->lookup(lookupName);
-      if (traitSym != nullptr) {
-        result = std::make_unique<Value>(*traitSym);
-      } else {
-        result = std::make_unique<Value>(typ);
-      }
-      return;
-    }
-    if (sym == nullptr) {
-      throw CodegenError(node->getSpan(), "Type not found: {}", node->getName());
-    }
-    if (!explicitTypeArgs.empty()) {
-      const Class* templateClass = nullptr;
-      if (auto gitClass = genericClasses.find(lookupName); gitClass != genericClasses.end()) {
-        templateClass = gitClass->second;
-      } else if (sym != nullptr && sym->getGenericClassTemplate() != nullptr) {
-        templateClass = static_cast<const Class*>(sym->getGenericClassTemplate());
-      }
-      if (templateClass == nullptr) {
-        throw CodegenError(node->getSpan(), "Type {} is not a generic class", node->getName());
-      }
-      sym = specializeClass(templateClass, {}, explicitTypeArgs);
-      typ = sym->getType();
-    }
-    if (sym->getType()->getLlvmType() == nullptr) {
-      getOrCreateLlvmType(sym->getType());
-    }
-
-    result = std::make_unique<Value>(*sym);
-  } else {
-    throw CodegenError(node->getSpan(), "Unimplemented type {}", NAMEOF_ENUM(node->getType()));
-  }
-}
-
-auto Codegen::getOrCreateListStructType(lesma::Type* listType) -> llvm::StructType* {
-  std::string typeName = "lesma.list";
-  if (listType != nullptr && listType->getElementType() != nullptr) {
-    getOrCreateLlvmType(listType->getElementType());
-    typeName += "." + MangleUtils::getTypeMangledName({}, listType->getElementType());
-  }
-  auto existing = listStructTypes.find(typeName);
-  if (existing != listStructTypes.end()) {
-    return existing->second;
-  }
-  auto* structType = llvm::StructType::create(
-      theModule->getContext(), {builder->getPtrTy(), builder->getInt64Ty(), builder->getInt64Ty()},
-      typeName, false);
-  listStructTypes[typeName] = structType;
-  return structType;
-}
-
-auto Codegen::getListStoredElementType(lesma::Type* listType) -> llvm::Type* {
-  auto* elementType = listType->getElementType();
-  if (elementType != nullptr && elementType->is(BaseType::TY_CLASS)) {
-    return builder->getPtrTy();
-  }
-  return getOrCreateLlvmType(elementType);
-}
-
-auto Codegen::getListStoredElementValue(llvm::SMRange span, lesma::Value* value,
-                                        lesma::Type* elementType) -> llvm::Value* {
-  if (elementType != nullptr && elementType->is(BaseType::TY_CLASS)) {
-    return value->getLlvmValue();
-  }
-  auto castVal = cast(span, value, elementType);
-  return castVal->getLlvmValue();
-}
-
-auto Codegen::emitCalloc(llvm::Value* count, llvm::Value* size, const llvm::Twine& name)
-    -> llvm::Value* {
-  auto callocFn = theModule->getOrInsertFunction(
-      "calloc", llvm::FunctionType::get(builder->getPtrTy(),
-                                        {builder->getInt64Ty(), builder->getInt64Ty()}, false));
-  return builder->CreateCall(callocFn, {count, size}, name);
-}
-
-auto Codegen::emitMalloc(llvm::Value* size, const llvm::Twine& name) -> llvm::Value* {
-  auto mallocFn = theModule->getOrInsertFunction(
-      "malloc", llvm::FunctionType::get(builder->getPtrTy(), {builder->getInt64Ty()}, false));
-  return builder->CreateCall(mallocFn, {size}, name);
-}
-
-auto Codegen::emitRealloc(llvm::Value* ptr, llvm::Value* size, const llvm::Twine& name)
-    -> llvm::Value* {
-  auto reallocFn = theModule->getOrInsertFunction(
-      "realloc", llvm::FunctionType::get(builder->getPtrTy(),
-                                         {builder->getPtrTy(), builder->getInt64Ty()}, false));
-  return builder->CreateCall(reallocFn, {ptr, size}, name);
-}
-
-auto Codegen::emitFree(llvm::Value* ptr) -> void {
-  auto freeFn = theModule->getOrInsertFunction(
-      "free", llvm::FunctionType::get(builder->getVoidTy(), {builder->getPtrTy()}, false));
-  builder->CreateCall(freeFn, {ptr});
-}
-
-auto Codegen::emitExit(int code) -> void {
-  auto exitFn = theModule->getOrInsertFunction(
-      "exit", llvm::FunctionType::get(builder->getVoidTy(), {builder->getInt64Ty()}, false));
-  builder->CreateCall(exitFn, {builder->getInt64(code)});
-}
-
-auto Codegen::emitListLength(lesma::Type* listType, llvm::Value* listHandle) -> llvm::Value* {
-  auto* structType = getOrCreateListStructType(listType);
-  auto* lenPtr = builder->CreateStructGEP(structType, listHandle, 1, "list.len.ptr");
-  return builder->CreateLoad(builder->getInt64Ty(), lenPtr, "list.len");
-}
-
-auto Codegen::emitListCapacity(lesma::Type* listType, llvm::Value* listHandle) -> llvm::Value* {
-  auto* structType = getOrCreateListStructType(listType);
-  auto* capPtr = builder->CreateStructGEP(structType, listHandle, 2, "list.cap.ptr");
-  return builder->CreateLoad(builder->getInt64Ty(), capPtr, "list.cap");
-}
-
-auto Codegen::emitListDataPtr(lesma::Type* listType, llvm::Value* listHandle) -> llvm::Value* {
-  auto* structType = getOrCreateListStructType(listType);
-  auto* dataPtrPtr = builder->CreateStructGEP(structType, listHandle, 0, "list.data.ptr.ptr");
-  return builder->CreateLoad(builder->getPtrTy(), dataPtrPtr, "list.data.ptr");
-}
-
-auto Codegen::emitStoreListDataPtr(lesma::Type* listType, llvm::Value* listHandle,
-                                   llvm::Value* dataValue) -> void {
-  auto* structType = getOrCreateListStructType(listType);
-  auto* slot = builder->CreateStructGEP(structType, listHandle, 0, "list.data.slot");
-  builder->CreateStore(dataValue, slot);
-}
-
-auto Codegen::emitStoreListLength(lesma::Type* listType, llvm::Value* listHandle,
-                                  llvm::Value* length) -> void {
-  auto* structType = getOrCreateListStructType(listType);
-  auto* slot = builder->CreateStructGEP(structType, listHandle, 1, "list.len.slot");
-  builder->CreateStore(length, slot);
-}
-
-auto Codegen::emitStoreListCapacity(lesma::Type* listType, llvm::Value* listHandle,
-                                    llvm::Value* capacity) -> void {
-  auto* structType = getOrCreateListStructType(listType);
-  auto* slot = builder->CreateStructGEP(structType, listHandle, 2, "list.cap.slot");
-  builder->CreateStore(capacity, slot);
-}
-
-auto Codegen::emitListBoundsCheck(llvm::SMRange span, lesma::Type* listType,
-                                  llvm::Value* listHandle, llvm::Value* index) -> void {
-  (void) span;
-  llvm::Function* parentFunction = builder->GetInsertBlock()->getParent();
-  auto* okBlock =
-      llvm::BasicBlock::Create(theModule->getContext(), "list.bounds.ok", parentFunction);
-  auto* failBlock =
-      llvm::BasicBlock::Create(theModule->getContext(), "list.bounds.fail", parentFunction);
-  auto* length = emitListLength(listType, listHandle);
-  auto* nonNegative = builder->CreateICmpSGE(index, builder->getInt64(0));
-  auto* inRange = builder->CreateICmpSLT(index, length);
-  builder->CreateCondBr(builder->CreateLogicalAnd(nonNegative, inRange), okBlock, failBlock);
-
-  builder->SetInsertPoint(failBlock);
-  emitExit(1);
-  builder->CreateUnreachable();
-
-  builder->SetInsertPoint(okBlock);
-}
-
-auto Codegen::emitListElementPointer(llvm::SMRange span, lesma::Type* listType,
-                                     llvm::Value* listHandle, llvm::Value* index) -> llvm::Value* {
-  emitListBoundsCheck(span, listType, listHandle, index);
-  auto* dataPtr = emitListDataPtr(listType, listHandle);
-  return builder->CreateGEP(getListStoredElementType(listType), dataPtr, index, "list.elem.ptr");
-}
-
-auto Codegen::emitListEnsureCapacity(lesma::Type* listType, llvm::Value* listHandle,
-                                     llvm::Value* minCapacity) -> void {
-  llvm::Function* parentFunction = builder->GetInsertBlock()->getParent();
-  auto* currentBlock = builder->GetInsertBlock();
-  auto* growBlock = llvm::BasicBlock::Create(theModule->getContext(), "list.grow", parentFunction);
-  auto* doneBlock =
-      llvm::BasicBlock::Create(theModule->getContext(), "list.grow.done", parentFunction);
-  auto* capacity = emitListCapacity(listType, listHandle);
-  auto* currentData = emitListDataPtr(listType, listHandle);
-  builder->CreateCondBr(builder->CreateICmpSGE(capacity, minCapacity), doneBlock, growBlock);
-
-  builder->SetInsertPoint(growBlock);
-  auto* doubledCap = builder->CreateMul(capacity, builder->getInt64(2), "list.cap.doubled");
-  auto* baseCap = builder->CreateSelect(builder->CreateICmpEQ(capacity, builder->getInt64(0)),
-                                        builder->getInt64(1), doubledCap);
-  auto* newCap = builder->CreateSelect(builder->CreateICmpSGE(baseCap, minCapacity), baseCap,
-                                       minCapacity, "list.cap.new");
-  auto* elementSize = builder->getInt64(theModule->getDataLayout()
-                                            .getTypeAllocSize(getListStoredElementType(listType))
-                                            .getFixedValue());
-  auto* newBytes = builder->CreateMul(newCap, elementSize, "list.grow.bytes");
-  auto* allocBlock =
-      llvm::BasicBlock::Create(theModule->getContext(), "list.grow.alloc", parentFunction);
-  auto* reallocBlock =
-      llvm::BasicBlock::Create(theModule->getContext(), "list.grow.realloc", parentFunction);
-  builder->CreateCondBr(
-      builder->CreateICmpEQ(currentData, llvm::ConstantPointerNull::get(builder->getPtrTy())),
-      allocBlock, reallocBlock);
-
-  builder->SetInsertPoint(allocBlock);
-  auto* allocatedData = emitMalloc(newBytes, "list.grow.alloc");
-  builder->CreateBr(doneBlock);
-
-  builder->SetInsertPoint(reallocBlock);
-  auto* reallocatedData = emitRealloc(currentData, newBytes, "list.grow.realloc");
-  builder->CreateBr(doneBlock);
-
-  builder->SetInsertPoint(doneBlock);
-  auto* newData = builder->CreatePHI(builder->getPtrTy(), 3, "list.grow.result");
-  auto* finalCapacity = builder->CreatePHI(builder->getInt64Ty(), 3, "list.grow.capacity");
-  newData->addIncoming(currentData, currentBlock);
-  newData->addIncoming(allocatedData, allocBlock);
-  newData->addIncoming(reallocatedData, reallocBlock);
-  finalCapacity->addIncoming(capacity, currentBlock);
-  finalCapacity->addIncoming(newCap, allocBlock);
-  finalCapacity->addIncoming(newCap, reallocBlock);
-  emitStoreListDataPtr(listType, listHandle, newData);
-  emitStoreListCapacity(listType, listHandle, finalCapacity);
-}
-
-auto Codegen::emitListDeepCopy(lesma::Type* listType, llvm::Value* listHandle) -> llvm::Value* {
-  auto* structType = getOrCreateListStructType(listType);
-  auto* headerSize =
-      builder->getInt64(theModule->getDataLayout().getTypeAllocSize(structType).getFixedValue());
-  auto* newHandle = emitMalloc(headerSize, "list.copy.header");
-  auto* length = emitListLength(listType, listHandle);
-  auto* capacity = emitListCapacity(listType, listHandle);
-  emitStoreListLength(listType, newHandle, length);
-  emitStoreListCapacity(listType, newHandle, capacity);
-  auto* elementType = listType->getElementType();
-  auto* elementLlvmType = getListStoredElementType(listType);
-  auto* elementSize = builder->getInt64(
-      theModule->getDataLayout().getTypeAllocSize(elementLlvmType).getFixedValue());
-  auto* hasCapacity = builder->CreateICmpSGT(capacity, builder->getInt64(0));
-  auto* dataSlot = builder->CreateAlloca(builder->getPtrTy(), nullptr, "list.copy.data.slot");
-  builder->CreateStore(llvm::ConstantPointerNull::get(builder->getPtrTy()), dataSlot);
-
-  llvm::Function* parentFunction = builder->GetInsertBlock()->getParent();
-  auto* copyElementsBlock =
-      llvm::BasicBlock::Create(theModule->getContext(), "list.copy.elements", parentFunction);
-  auto* doneBlock =
-      llvm::BasicBlock::Create(theModule->getContext(), "list.copy.done", parentFunction);
-  builder->CreateCondBr(hasCapacity, copyElementsBlock, doneBlock);
-
-  builder->SetInsertPoint(copyElementsBlock);
-  auto* newBytes = builder->CreateMul(capacity, elementSize, "list.copy.bytes");
-  auto* initBytes = builder->CreateMul(length, elementSize, "list.copy.initbytes");
-  auto* newData = emitMalloc(newBytes, "list.copy.data");
-  builder->CreateStore(newData, dataSlot);
-  auto isNestedListLike = [elementType]() -> bool {
-    if (elementType == nullptr || !elementType->is(BaseType::TY_CLASS)) {
-      return false;
-    }
-    auto fields = elementType->getFields();
-    return !fields.empty() && fields.front()->type != nullptr &&
-           fields.front()->type->is(BaseType::TY_ARRAY);
-  };
-  if (elementType->is(BaseType::TY_ARRAY) || isNestedListLike()) {
-    auto* indexPtr = builder->CreateAlloca(builder->getInt64Ty(), nullptr, "list.copy.index");
-    builder->CreateStore(builder->getInt64(0), indexPtr);
-    auto* loopCond =
-        llvm::BasicBlock::Create(theModule->getContext(), "list.copy.loop.cond", parentFunction);
-    auto* loopBody =
-        llvm::BasicBlock::Create(theModule->getContext(), "list.copy.loop.body", parentFunction);
-    auto* loopInc =
-        llvm::BasicBlock::Create(theModule->getContext(), "list.copy.loop.inc", parentFunction);
-    builder->CreateBr(loopCond);
-
-    builder->SetInsertPoint(loopCond);
-    auto* index = builder->CreateLoad(builder->getInt64Ty(), indexPtr);
-    builder->CreateCondBr(builder->CreateICmpSLT(index, length), loopBody, doneBlock);
-
-    builder->SetInsertPoint(loopBody);
-    auto* oldElementPtr = emitListElementPointer({}, listType, listHandle, index);
-    auto* oldElement = builder->CreateLoad(elementLlvmType, oldElementPtr);
-    llvm::Value* copiedElement = nullptr;
-    if (elementType->is(BaseType::TY_ARRAY)) {
-      copiedElement = emitListDeepCopy(elementType, oldElement);
-    } else {
-      lesma::Value oldValue("", elementType, oldElement);
-      auto copiedValue = callMethodByName({}, &oldValue, "copy");
-      copiedElement = copiedValue->getLlvmValue();
-    }
-    auto* newElementPtr = builder->CreateGEP(elementLlvmType, newData, index, "list.copy.elem.ptr");
-    builder->CreateStore(copiedElement, newElementPtr);
-    builder->CreateBr(loopInc);
-
-    builder->SetInsertPoint(loopInc);
-    auto* nextIndex = builder->CreateAdd(builder->CreateLoad(builder->getInt64Ty(), indexPtr),
-                                         builder->getInt64(1));
-    builder->CreateStore(nextIndex, indexPtr);
-    builder->CreateBr(loopCond);
-  } else {
-    auto memcpyFn = theModule->getOrInsertFunction(
-        "memcpy", llvm::FunctionType::get(
-                      builder->getPtrTy(),
-                      {builder->getPtrTy(), builder->getPtrTy(), builder->getInt64Ty()}, false));
-    builder->CreateCall(memcpyFn, {newData, emitListDataPtr(listType, listHandle), initBytes});
-    builder->CreateBr(doneBlock);
-  }
-
-  builder->SetInsertPoint(doneBlock);
-  emitStoreListDataPtr(listType, newHandle, builder->CreateLoad(builder->getPtrTy(), dataSlot));
-  return newHandle;
-}
-
-auto Codegen::isListIntrinsicName(const std::string& functionName) const -> bool {
-  return functionName == "__list_len" || functionName == "__list_push" ||
-         functionName == "__list_pop" || functionName == "__list_clear" ||
-         functionName == "__list_copy" || functionName == "__list_get" ||
-         functionName == "__list_set" || functionName == "__buffer_new" ||
-         functionName == "__buffer_len" || functionName == "__buffer_push" ||
-         functionName == "__buffer_pop" || functionName == "__buffer_clear" ||
-         functionName == "__buffer_copy" || functionName == "__buffer_get" ||
-         functionName == "__buffer_set" || functionName == "__cstr_byte_at" ||
-         functionName == "__cstr_byte_set" || functionName == "__str_concat" ||
-         functionName == "__str_slice" || functionName == "__cstr_index_of" ||
-         functionName == "__cstr_offset";
-}
-
-auto Codegen::isBuiltinListBuiltinMethodName(const std::string& methodName) const -> bool {
-  return methodName == "len" || methodName == "clear" || methodName == "push" ||
-         methodName == "pop" || methodName == "copy" ||
-         methodName == std::string{OperatorUtils::SUBSCRIPT_GET_NAME} ||
-         methodName == std::string{OperatorUtils::SUBSCRIPT_SET_NAME};
-}
-
-auto Codegen::genListIntrinsicCall(const FuncCall* node,
-                                   const std::vector<lesma::Type*>& paramTypes,
-                                   const std::vector<llvm::Value*>& paramsLLVM)
-    -> std::unique_ptr<lesma::Value> {
-  if (node->getName() == "__cstr_byte_at") {
-    if (paramTypes.size() != 2U || paramsLLVM.size() != 2U) {
-      throw CodegenError(node->getSpan(), "__cstr_byte_at expects cstr and index");
-    }
-    llvm::Type* i8 = llvm::Type::getInt8Ty(theModule->getContext());
-    auto* gep = builder->CreateInBoundsGEP(i8, paramsLLVM[0], paramsLLVM[1], "cstr.byte.ptr");
-    auto* byteVal = builder->CreateLoad(i8, gep);
-    auto* int8Ty = cacheType(std::make_unique<Type>(BaseType::TY_INT, i8));
-    int8Ty->setIntWidth(8);
-    return std::make_unique<Value>("", int8Ty, byteVal);
-  }
-  if (node->getName() == "__cstr_byte_set") {
-    if (paramTypes.size() != 3U || paramsLLVM.size() != 3U) {
-      throw CodegenError(node->getSpan(), "__cstr_byte_set expects cstr, index, and int8");
-    }
-    llvm::Type* i8 = llvm::Type::getInt8Ty(theModule->getContext());
-    auto* gep = builder->CreateInBoundsGEP(i8, paramsLLVM[0], paramsLLVM[1], "cstr.set.ptr");
-    llvm::Value* v = paramsLLVM[2];
-    if (!v->getType()->isIntegerTy(8)) {
-      v = builder->CreateTrunc(v, i8);
-    }
-    builder->CreateStore(v, gep);
-    return std::make_unique<Value>(
-        "", cacheType(std::make_unique<Type>(BaseType::TY_VOID, builder->getVoidTy())), nullptr);
-  }
-  if (node->getName() == "__str_concat") {
-    if (paramTypes.size() != 2U || paramsLLVM.size() != 2U) {
-      throw CodegenError(node->getSpan(), "__str_concat expects two cstr arguments");
-    }
-    llvm::Type* i8 = llvm::Type::getInt8Ty(theModule->getContext());
-    auto strlenFn = theModule->getOrInsertFunction(
-        "strlen", llvm::FunctionType::get(builder->getInt64Ty(), {builder->getPtrTy()}, false));
-    auto memcpyFn = theModule->getOrInsertFunction(
-        "memcpy", llvm::FunctionType::get(
-                      builder->getPtrTy(),
-                      {builder->getPtrTy(), builder->getPtrTy(), builder->getInt64Ty()}, false));
-    llvm::Value* la = paramsLLVM[0];
-    llvm::Value* lb = paramsLLVM[1];
-    auto* lenA = builder->CreateCall(strlenFn, {la}, "strcat.lenA");
-    auto* lenB = builder->CreateCall(strlenFn, {lb}, "strcat.lenB");
-    auto* total = builder->CreateAdd(
-        builder->CreateAdd(lenA, lenB), builder->getInt64(1), "strcat.total");
-    llvm::Value* buf = emitMalloc(total, "strcat.buf");
-    builder->CreateCall(memcpyFn, {buf, la, lenA});
-    auto* tail = builder->CreateInBoundsGEP(i8, buf, lenA, "strcat.tail");
-    builder->CreateCall(memcpyFn, {tail, lb, lenB});
-    auto* endPtr = builder->CreateInBoundsGEP(
-        i8, buf, builder->CreateSub(total, builder->getInt64(1)), "strcat.nul");
-    builder->CreateStore(llvm::ConstantInt::get(i8, 0), endPtr);
-    auto* cstrTy = cacheType(std::make_unique<Type>(BaseType::TY_STRING, builder->getPtrTy()));
-    return std::make_unique<Value>("", cstrTy, buf);
-  }
-  if (node->getName() == "__str_slice") {
-    if (paramTypes.size() != 3U || paramsLLVM.size() != 3U) {
-      throw CodegenError(node->getSpan(), "__str_slice expects cstr, start, and length");
-    }
-    llvm::Type* i8 = llvm::Type::getInt8Ty(theModule->getContext());
-    auto memcpyFn = theModule->getOrInsertFunction(
-        "memcpy", llvm::FunctionType::get(
-                      builder->getPtrTy(),
-                      {builder->getPtrTy(), builder->getPtrTy(), builder->getInt64Ty()}, false));
-    llvm::Value* s = paramsLLVM[0];
-    llvm::Value* start = paramsLLVM[1];
-    llvm::Value* len = paramsLLVM[2];
-    llvm::Function* parentFunction = builder->GetInsertBlock()->getParent();
-    auto* emptyBlock = llvm::BasicBlock::Create(theModule->getContext(), "slice.empty", parentFunction);
-    auto* copyBlock = llvm::BasicBlock::Create(theModule->getContext(), "slice.copy", parentFunction);
-    auto* mergeBlock = llvm::BasicBlock::Create(theModule->getContext(), "slice.merge", parentFunction);
-    auto* lenPos = builder->CreateICmpSGT(len, builder->getInt64(0));
-    builder->CreateCondBr(lenPos, copyBlock, emptyBlock);
-
-    builder->SetInsertPoint(emptyBlock);
-    llvm::Value* emptyBuf = emitMalloc(builder->getInt64(1), "slice.empty.buf");
-    builder->CreateStore(llvm::ConstantInt::get(i8, 0), emptyBuf);
-    builder->CreateBr(mergeBlock);
-
-    builder->SetInsertPoint(copyBlock);
-    auto* total = builder->CreateAdd(len, builder->getInt64(1), "slice.total");
-    llvm::Value* buf = emitMalloc(total, "slice.buf");
-    auto* src = builder->CreateInBoundsGEP(i8, s, start, "slice.src");
-    builder->CreateCall(memcpyFn, {buf, src, len});
-    auto* nulPtr = builder->CreateInBoundsGEP(i8, buf, len, "slice.nul");
-    builder->CreateStore(llvm::ConstantInt::get(i8, 0), nulPtr);
-    builder->CreateBr(mergeBlock);
-
-    builder->SetInsertPoint(mergeBlock);
-    auto* phi = builder->CreatePHI(builder->getPtrTy(), 2, "slice.result");
-    phi->addIncoming(emptyBuf, emptyBlock);
-    phi->addIncoming(buf, copyBlock);
-    auto* cstrTy = cacheType(std::make_unique<Type>(BaseType::TY_STRING, builder->getPtrTy()));
-    return std::make_unique<Value>("", cstrTy, phi);
-  }
-  if (node->getName() == "__cstr_index_of") {
-    if (paramTypes.size() != 2U || paramsLLVM.size() != 2U) {
-      throw CodegenError(node->getSpan(), "__cstr_index_of expects haystack and needle");
-    }
-    auto strstrFn = theModule->getOrInsertFunction(
-        "strstr", llvm::FunctionType::get(builder->getPtrTy(),
-                                          {builder->getPtrTy(), builder->getPtrTy()}, false));
-    llvm::Value* hay = paramsLLVM[0];
-    llvm::Value* needle = paramsLLVM[1];
-    llvm::Value* found = builder->CreateCall(strstrFn, {hay, needle}, "idx.found");
-    llvm::Function* parentFunction = builder->GetInsertBlock()->getParent();
-    auto* okBlock = llvm::BasicBlock::Create(theModule->getContext(), "idx.ok", parentFunction);
-    auto* failBlock = llvm::BasicBlock::Create(theModule->getContext(), "idx.fail", parentFunction);
-    auto* mergeBlock = llvm::BasicBlock::Create(theModule->getContext(), "idx.merge", parentFunction);
-    builder->CreateCondBr(
-        builder->CreateICmpEQ(found, llvm::ConstantPointerNull::get(builder->getPtrTy())), failBlock,
-        okBlock);
-
-    builder->SetInsertPoint(failBlock);
-    builder->CreateBr(mergeBlock);
-
-    builder->SetInsertPoint(okBlock);
-    auto* diff = builder->CreateSub(builder->CreatePtrToInt(found, builder->getInt64Ty()),
-                                    builder->CreatePtrToInt(hay, builder->getInt64Ty()), "idx.diff");
-    builder->CreateBr(mergeBlock);
-
-    builder->SetInsertPoint(mergeBlock);
-    auto* phi = builder->CreatePHI(builder->getInt64Ty(), 2, "idx.result");
-    phi->addIncoming(builder->getInt64(-1), failBlock);
-    phi->addIncoming(diff, okBlock);
-    auto* intTy = cacheType(std::make_unique<Type>(BaseType::TY_INT, builder->getInt64Ty()));
-    intTy->setIntWidth(64);
-    return std::make_unique<Value>("", intTy, phi);
-  }
-  if (node->getName() == "__cstr_offset") {
-    if (paramTypes.size() != 2U || paramsLLVM.size() != 2U) {
-      throw CodegenError(node->getSpan(), "__cstr_offset expects cstr and offset");
-    }
-    llvm::Type* i8 = llvm::Type::getInt8Ty(theModule->getContext());
-    auto* out = builder->CreateInBoundsGEP(i8, paramsLLVM[0], paramsLLVM[1], "cstr.off");
-    auto* cstrTy = cacheType(std::make_unique<Type>(BaseType::TY_STRING, builder->getPtrTy()));
-    return std::make_unique<Value>("", cstrTy, out);
-  }
-  if (node->getName() == "__buffer_new") {
-    auto explicitTypeArgs = node->getExplicitTypeArgs();
-    if (explicitTypeArgs.size() != 1U) {
-      throw CodegenError(node->getSpan(), "__buffer_new<T>() expects exactly one type argument");
-    }
-    explicitTypeArgs.front()->accept(*this);
-    auto* listType =
-        cacheType(std::make_unique<Type>(BaseType::TY_ARRAY, nullptr, result->getType()));
-    listType->setDisplayName("__buffer");
-    getOrCreateLlvmType(listType);
-    auto* listStructTy = getOrCreateListStructType(listType);
-    auto* headerSize = builder->getInt64(
-        theModule->getDataLayout().getTypeAllocSize(listStructTy).getFixedValue());
-    auto* listHandle = emitMalloc(headerSize, "buffer.header");
-    emitStoreListDataPtr(listType, listHandle, llvm::ConstantPointerNull::get(builder->getPtrTy()));
-    emitStoreListLength(listType, listHandle, builder->getInt64(0));
-    emitStoreListCapacity(listType, listHandle, builder->getInt64(0));
-    return std::make_unique<Value>("", listType, listHandle);
-  }
-  if (paramTypes.empty() || paramsLLVM.empty()) {
-    throw CodegenError(node->getSpan(), "List intrinsic {} requires a list argument",
-                       node->getName());
-  }
-  auto* listType = paramTypes.front();
-  auto* listHandle = paramsLLVM.front();
-  if (listType == nullptr || !listType->is(BaseType::TY_ARRAY) ||
-      listType->getElementType() == nullptr) {
-    throw CodegenError(node->getSpan(), "List intrinsic {} requires list<T>", node->getName());
-  }
-
-  if (node->getName() == "__list_len" || node->getName() == "__buffer_len") {
-    auto* lenTy = cacheType(std::make_unique<Type>(BaseType::TY_INT, builder->getInt64Ty()));
-    lenTy->setIntWidth(64);
-    return std::make_unique<Value>("", lenTy, emitListLength(listType, listHandle));
-  }
-  if (node->getName() == "__list_copy" || node->getName() == "__buffer_copy") {
-    return std::make_unique<Value>("", listType, emitListDeepCopy(listType, listHandle));
-  }
-  if (node->getName() == "__list_clear" || node->getName() == "__buffer_clear") {
-    auto* currentData = emitListDataPtr(listType, listHandle);
-    llvm::Function* parentFunction = builder->GetInsertBlock()->getParent();
-    auto* freeBlock =
-        llvm::BasicBlock::Create(theModule->getContext(), "list.clear.free", parentFunction);
-    auto* doneBlock =
-        llvm::BasicBlock::Create(theModule->getContext(), "list.clear.done", parentFunction);
-    builder->CreateCondBr(
-        builder->CreateICmpNE(currentData, llvm::ConstantPointerNull::get(builder->getPtrTy())),
-        freeBlock, doneBlock);
-    builder->SetInsertPoint(freeBlock);
-    emitFree(currentData);
-    builder->CreateBr(doneBlock);
-    builder->SetInsertPoint(doneBlock);
-    emitStoreListDataPtr(listType, listHandle, llvm::ConstantPointerNull::get(builder->getPtrTy()));
-    emitStoreListLength(listType, listHandle, builder->getInt64(0));
-    emitStoreListCapacity(listType, listHandle, builder->getInt64(0));
-    return std::make_unique<Value>(
-        "", cacheType(std::make_unique<Type>(BaseType::TY_VOID, builder->getVoidTy())), nullptr);
-  }
-  if (node->getName() == "__list_push" || node->getName() == "__buffer_push") {
-    if (paramTypes.size() != 2U || paramsLLVM.size() != 2U) {
-      throw CodegenError(node->getSpan(), "__list_push expects list and value");
-    }
-    auto* length = emitListLength(listType, listHandle);
-    auto* nextLength = builder->CreateAdd(length, builder->getInt64(1));
-    emitListEnsureCapacity(listType, listHandle, nextLength);
-    auto* elementPtr =
-        builder->CreateGEP(getListStoredElementType(listType),
-                           emitListDataPtr(listType, listHandle), length, "list.push.ptr");
-    builder->CreateStore(paramsLLVM[1], elementPtr);
-    emitStoreListLength(listType, listHandle, nextLength);
-    return std::make_unique<Value>(
-        "", cacheType(std::make_unique<Type>(BaseType::TY_VOID, builder->getVoidTy())), nullptr);
-  }
-  if (node->getName() == "__list_get" || node->getName() == "__buffer_get") {
-    if (paramTypes.size() != 2U || paramsLLVM.size() != 2U) {
-      throw CodegenError(node->getSpan(), "__list_get expects list and index");
-    }
-    auto* elementPtr = emitListElementPointer(node->getSpan(), listType, listHandle, paramsLLVM[1]);
-    return std::make_unique<Value>(
-        "", listType->getElementType(),
-        builder->CreateLoad(getListStoredElementType(listType), elementPtr));
-  }
-  if (node->getName() == "__list_set" || node->getName() == "__buffer_set") {
-    if (paramTypes.size() != 3U || paramsLLVM.size() != 3U) {
-      throw CodegenError(node->getSpan(), "__list_set expects list, index, and value");
-    }
-    auto* elementPtr = emitListElementPointer(node->getSpan(), listType, listHandle, paramsLLVM[1]);
-    builder->CreateStore(paramsLLVM[2], elementPtr);
-    return std::make_unique<Value>(
-        "", cacheType(std::make_unique<Type>(BaseType::TY_VOID, builder->getVoidTy())), nullptr);
-  }
-  if (node->getName() == "__list_pop" || node->getName() == "__buffer_pop") {
-    auto* length = emitListLength(listType, listHandle);
-    emitListBoundsCheck(node->getSpan(), listType, listHandle,
-                        builder->CreateSub(length, builder->getInt64(1)));
-    auto* newLength = builder->CreateSub(length, builder->getInt64(1), "list.pop.len");
-    auto* elementPtr =
-        builder->CreateGEP(getListStoredElementType(listType),
-                           emitListDataPtr(listType, listHandle), newLength, "list.pop.ptr");
-    auto* poppedValue = builder->CreateLoad(getListStoredElementType(listType), elementPtr);
-    emitStoreListLength(listType, listHandle, newLength);
-    return std::make_unique<Value>("", listType->getElementType(), poppedValue);
-  }
-
-  throw CodegenError(node->getSpan(), "Unknown list intrinsic {}", node->getName());
-}
-
-namespace {
-[[nodiscard]] auto stripSpacesCopy(std::string s) -> std::string {
-  s.erase(std::remove(s.begin(), s.end(), ' '), s.end());
-  return s;
-}
-[[nodiscard]] auto classDisplayNamesMatch(const std::string& a, const std::string& b) -> bool {
-  if (a == b) {
-    return true;
-  }
-  if (a.size() < 6 || b.size() < 6 || a.compare(0, 5, "list<") != 0 || b.compare(0, 5, "list<") != 0) {
-    return false;
-  }
-  return stripSpacesCopy(a) == stripSpacesCopy(b);
-}
-} // namespace
-
-auto Codegen::tryEnsureStdlibListClassSpecialized(lesma::Type* classTy) -> void {
-  if (classTy == nullptr || !classTy->is(BaseType::TY_CLASS)) {
-    return;
-  }
-  const std::string& dn = classTy->getDisplayName();
-  if (dn.size() < 6 || dn.compare(0, 5, "list<") != 0 || dn.back() != '>') {
-    return;
-  }
-  auto git = genericClasses.find("list");
-  if (git == genericClasses.end() || git->second == nullptr) {
-    return;
-  }
-  lesma::Type* typeArg = nullptr;
-  if (auto envIt = specializedClassTypeEnvs.find(classTy); envIt != specializedClassTypeEnvs.end()) {
-    if (auto tIt = envIt->second.find("T"); tIt != envIt->second.end()) {
-      typeArg = tIt->second;
-    }
-  }
-  if (typeArg == nullptr) {
-    for (const auto& [ty, env] : specializedClassTypeEnvs) {
-      if (ty != nullptr && classDisplayNamesMatch(ty->getDisplayName(), dn)) {
-        if (auto tIt = env.find("T"); tIt != env.end()) {
-          typeArg = tIt->second;
-          break;
-        }
-      }
-    }
-  }
-  if (typeArg == nullptr) {
-    auto fields = classTy->getFields();
-    if (!fields.empty() && fields[0]->type != nullptr && fields[0]->type->is(BaseType::TY_ARRAY) &&
-        fields[0]->type->getElementType() != nullptr) {
-      typeArg = fields[0]->type->getElementType();
-    }
-  }
-  if (typeArg == nullptr) {
-    return;
-  }
-  specializeClass(git->second, {}, {typeArg});
-}
-
-auto Codegen::lookupClassStructSymbol(lesma::Type* classTy) -> Value* {
-  if (classTy == nullptr || !classTy->is(BaseType::TY_CLASS)) {
-    return nullptr;
-  }
-  llvm::Type* lt = classTy->getLlvmType();
-  if (lt != nullptr) {
-    if (auto* st = llvm::dyn_cast<llvm::StructType>(lt)) {
-      if (st->hasName()) {
-        if (Value* v = scope->lookupStruct(st->getName().str())) {
-          return v;
-        }
-      }
-    }
-  }
-  if (!classTy->getDisplayName().empty()) {
-    if (Value* v = scope->lookupStruct(classTy->getDisplayName())) {
-      return v;
-    }
-    // Typechecker may use a different Type* than the one codegen registered when specializing
-    // generics; match the struct symbol by display name.
-    for (const auto& [ty, sym] : specializedClassSymbolsByType) {
-      if (ty != nullptr && classDisplayNamesMatch(ty->getDisplayName(), classTy->getDisplayName())) {
-        return sym;
-      }
-    }
-  }
-  tryEnsureStdlibListClassSpecialized(classTy);
-  if (!classTy->getDisplayName().empty()) {
-    if (Value* v = scope->lookupStruct(classTy->getDisplayName())) {
-      return v;
-    }
-    for (const auto& [ty, sym] : specializedClassSymbolsByType) {
-      if (ty != nullptr && classDisplayNamesMatch(ty->getDisplayName(), classTy->getDisplayName())) {
-        return sym;
-      }
-    }
-  }
-  lt = classTy->getLlvmType();
-  if (lt != nullptr) {
-    if (auto* st = llvm::dyn_cast<llvm::StructType>(lt)) {
-      if (st->hasName()) {
-        if (Value* v = scope->lookupStruct(st->getName().str())) {
-          return v;
-        }
-      }
-    }
-  }
-  return nullptr;
-}
-
-auto Codegen::getOrCreateLlvmType(lesma::Type* type) -> llvm::Type* {
-  if (type->getLlvmType() != nullptr) {
-    return type->getLlvmType();
-  }
-  switch (type->getBaseType()) {
-  case BaseType::TY_GENERIC: {
-    auto it = currentGenericTypes.find(type->getGenericName());
-    if (it == currentGenericTypes.end()) {
-      throw CodegenError({}, "Unknown generic type {}", type->getGenericName());
-    }
-    return getOrCreateLlvmType(it->second);
-  }
-  case BaseType::TY_INT: {
-    const unsigned w = type->getIntWidth();
-    llvm::Type* mapped = builder->getInt64Ty();
-    if (w == 8) {
-      mapped = builder->getInt8Ty();
-    } else if (w == 16) {
-      mapped = builder->getInt16Ty();
-    } else if (w == 32) {
-      mapped = builder->getInt32Ty();
-    }
-    type->setLlvmType(mapped);
-    break;
-  }
-  case BaseType::TY_FLOAT:
-    type->setLlvmType(builder->getDoubleTy());
-    break;
-  case BaseType::TY_BOOL:
-    type->setLlvmType(builder->getInt1Ty());
-    break;
-  case BaseType::TY_STRING:
-    type->setLlvmType(builder->getPtrTy());
-    break;
-  case BaseType::TY_VOID:
-    type->setLlvmType(builder->getVoidTy());
-    break;
-  case BaseType::TY_PTR:
-    if (type->getElementType() != nullptr) {
-      getOrCreateLlvmType(type->getElementType());
-    }
-    type->setLlvmType(builder->getPtrTy());
-    break;
-  case BaseType::TY_ARRAY: {
-    if (type->getElementType() != nullptr) {
-      getOrCreateLlvmType(type->getElementType());
-    }
-    getOrCreateListStructType(type);
-    type->setLlvmType(builder->getPtrTy());
-    break;
-  }
-  case BaseType::TY_FUNCTION:
-    getOrCreateLlvmType(type->getReturnType());
-    for (auto* f : type->getFields()) {
-      getOrCreateLlvmType(f->type);
-    }
-    type->setLlvmType(builder->getPtrTy());
-    break;
-  case BaseType::TY_TRAIT_EXISTENTIAL: {
-    llvm::Type* st =
-        llvm::StructType::get(theModule->getContext(), {builder->getPtrTy(), builder->getPtrTy()});
-    type->setLlvmType(st);
-    break;
-  }
-  case BaseType::TY_CLASS:
-  case BaseType::TY_ENUM: {
-    // Create opaque struct first to break recursion (e.g. class with field
-    // *Self). Use the Lesma display name so imported class types (e.g. stdlib
-    // str) get a stable LLVM struct name when this module never visits Class*.
-    llvm::StructType* st = nullptr;
-    if (!type->getDisplayName().empty()) {
-      st = llvm::StructType::create(theModule->getContext(), type->getDisplayName());
-    } else {
-      st = llvm::StructType::create(theModule->getContext());
-    }
-    type->setLlvmType(st);
-    std::vector<llvm::Type*> elementTypes;
-    for (auto* f : type->getFields()) {
-      elementTypes.push_back(getOrCreateLlvmType(f->type));
-    }
-    if (elementTypes.empty()) {
-      elementTypes.push_back(builder->getInt8Ty());
-    }
-    st->setBody(elementTypes);
-    break;
-  }
-  default:
-    type->setLlvmType(builder->getPtrTy());
-    break;
-  }
-  return type->getLlvmType();
-}
-
-auto Codegen::collectTraitMetadataFromAst() -> void {
-  Compound* ast = parser->getAst();
-  if (ast == nullptr) {
-    return;
-  }
-  for (Statement* stmt : ast->getChildren()) {
-    if (auto* tr = dynamic_cast<TraitDecl*>(stmt)) {
-      std::vector<std::string> order;
-      for (FuncDecl* req : tr->getRequirements()) {
-        order.push_back(req->getName());
-      }
-      traitRequirementMethodOrder[tr->getIdentifier()] = std::move(order);
-      traitDeclByName[tr->getIdentifier()] = tr;
-    }
-  }
-}
-
-auto Codegen::mergeImportedTraitMetadata(Codegen const& imported) -> void {
-  for (const auto& entry : imported.traitDeclByName) {
-    if (traitDeclByName.contains(entry.first)) {
-      continue;
-    }
-    traitDeclByName[entry.first] = entry.second;
-    auto ordIt = imported.traitRequirementMethodOrder.find(entry.first);
-    if (ordIt != imported.traitRequirementMethodOrder.end()) {
-      traitRequirementMethodOrder[entry.first] = ordIt->second;
-    }
-  }
-}
-
-auto Codegen::mergeImportedSpecializationState(Codegen const& imported) -> void {
-  for (const auto& entry : imported.genericClasses) {
-    genericClasses.insert(entry);
-  }
-  for (const auto& entry : imported.specializedClassTypeEnvs) {
-    specializedClassTypeEnvs.insert(entry);
-  }
-  for (const auto& entry : imported.specializationEnvs) {
-    specializationEnvs.insert(entry);
-  }
-}
-
-auto Codegen::findTraitRequirement(const TraitDecl* trait, const std::string& methodName) const
-    -> const FuncDecl* {
-  if (trait == nullptr) {
-    return nullptr;
-  }
-  for (FuncDecl* req : trait->getRequirements()) {
-    if (req->getName() == methodName) {
-      return req;
-    }
-  }
-  return nullptr;
-}
-
-auto Codegen::emitErasedThunkForTraitMethod(lesma::Type* classType, const std::string& traitName,
-                                            const FuncDecl* req) -> llvm::Function* {
-  std::string cacheKey = traitName + "|" + classType->getDisplayName() + "|" + req->getName();
-  if (auto it = traitThunkCache.find(cacheKey); it != traitThunkCache.end()) {
-    return it->second;
-  }
-
-  lesma::Type* selfPtr = cacheType(std::make_unique<Type>(BaseType::TY_PTR, nullptr, classType));
-  std::vector<lesma::Type*> lookupArgs = {selfPtr};
-  for (Parameter* p : req->getParameters()) {
-    if (p->type != nullptr) {
-      p->type->accept(*this);
-      lesma::Type* pt = result->getType();
-      if (pt->is(BaseType::TY_CLASS)) {
-        pt = cacheType(std::make_unique<Type>(BaseType::TY_PTR, nullptr, pt));
-      }
-      lookupArgs.push_back(pt);
-    }
-  }
-  Value* methodSym = scope->lookupFunction(req->getName(), lookupArgs);
-  if (methodSym == nullptr) {
-    for (const auto& importedScope : *importedScopes) {
-      if (importedScope == nullptr) {
-        continue;
-      }
-      methodSym = importedScope->lookupFunction(req->getName(), lookupArgs);
-      if (methodSym != nullptr) {
-        break;
-      }
-    }
-  }
-  if (methodSym == nullptr || methodSym->getLlvmValue() == nullptr) {
-    throw CodegenError(req->getSpan(), "Trait thunk: method {} not found for class {}",
-                       req->getName(), classType->getDisplayName());
-  }
-  auto* realFn = llvm::cast<llvm::Function>(methodSym->getLlvmValue());
-  llvm::FunctionType* rft = realFn->getFunctionType();
-
-  llvm::SmallVector<llvm::Type*, 8> tparams;
-  tparams.push_back(builder->getPtrTy());
-  for (unsigned i = 1; i < rft->getNumParams(); ++i) {
-    tparams.push_back(rft->getParamType(i));
-  }
-  llvm::FunctionType* tft = llvm::FunctionType::get(rft->getReturnType(), tparams, false);
-
-  std::string thunkName = "lesma.trait.thunk." + cacheKey;
-  for (char& c : thunkName) {
-    if (!(std::isalnum(static_cast<unsigned char>(c)) != 0 || c == '.' || c == '_')) {
-      c = '_';
-    }
-  }
-  llvm::Function* thunk = theModule->getFunction(thunkName);
-  if (thunk == nullptr) {
-    thunk = llvm::Function::Create(tft, llvm::Function::InternalLinkage, thunkName, *theModule);
-    llvm::BasicBlock* bb = llvm::BasicBlock::Create(theModule->getContext(), "entry", thunk);
-    llvm::IRBuilder<> tmpBuilder(bb);
-    llvm::Argument* rawArg = thunk->getArg(0);
-    llvm::Value* castSelf = tmpBuilder.CreateBitCast(rawArg, rft->getParamType(0));
-    llvm::SmallVector<llvm::Value*, 8> callArgs;
-    callArgs.push_back(castSelf);
-    for (unsigned i = 1; i < rft->getNumParams(); ++i) {
-      callArgs.push_back(thunk->getArg(i));
-    }
-    if (rft->getReturnType()->isVoidTy()) {
-      tmpBuilder.CreateCall(rft, realFn, callArgs);
-      tmpBuilder.CreateRetVoid();
-    } else {
-      llvm::Value* r = tmpBuilder.CreateCall(rft, realFn, callArgs);
-      tmpBuilder.CreateRet(r);
-    }
-  }
-  traitThunkCache[cacheKey] = thunk;
-  return thunk;
-}
-
-auto Codegen::getOrEmitWitnessTable(lesma::Type* classType, const std::string& traitName)
-    -> llvm::GlobalVariable* {
-  std::string cacheKey = traitName + "|" + classType->getDisplayName();
-  if (auto it = witnessGlobalCache.find(cacheKey); it != witnessGlobalCache.end()) {
-    return it->second;
-  }
-  const TraitDecl* tr = traitDeclByName[traitName];
-  if (tr == nullptr) {
-    throw CodegenError({}, "Codegen: unknown trait {}", traitName);
-  }
-  auto ordIt = traitRequirementMethodOrder.find(traitName);
-  if (ordIt == traitRequirementMethodOrder.end()) {
-    throw CodegenError({}, "Codegen: trait {} has no requirement order", traitName);
-  }
-  std::vector<llvm::Constant*> constants;
-  constants.reserve(ordIt->second.size());
-  for (const std::string& mn : ordIt->second) {
-    const FuncDecl* req = findTraitRequirement(tr, mn);
-    if (req == nullptr) {
-      throw CodegenError({}, "Codegen: trait {} missing requirement {}", traitName, mn);
-    }
-    llvm::Function* thunk = emitErasedThunkForTraitMethod(classType, traitName, req);
-    constants.push_back(llvm::ConstantExpr::getBitCast(thunk, builder->getPtrTy()));
-  }
-  llvm::ArrayType* at = llvm::ArrayType::get(builder->getPtrTy(), constants.size());
-  llvm::Constant* init = llvm::ConstantArray::get(at, constants);
-  std::string gname = "lesma.witness." + cacheKey;
-  for (char& c : gname) {
-    if (!(std::isalnum(static_cast<unsigned char>(c)) != 0 || c == '.' || c == '_')) {
-      c = '_';
-    }
-  }
-  auto* gv = new llvm::GlobalVariable(*theModule, at, true, llvm::GlobalValue::PrivateLinkage, init,
-                                      gname);
-  witnessGlobalCache[cacheKey] = gv;
-  return gv;
-}
-
-auto Codegen::emitBoxClassToExistential(lesma::Type* existentialType,
-                                        lesma::Type* classPtrLesmaType, llvm::Value* classPtrVal)
-    -> llvm::Value* {
-  lesma::Type* cls = classPtrLesmaType;
-  if (cls->is(BaseType::TY_PTR) && cls->getElementType() != nullptr &&
-      cls->getElementType()->is(BaseType::TY_CLASS)) {
-    cls = cls->getElementType();
-  }
-  if (!cls->is(BaseType::TY_CLASS)) {
-    throw CodegenError({}, "Box to existential requires class pointer");
-  }
-  const std::string traitName = existentialType->getDisplayName();
-  llvm::GlobalVariable* wit = getOrEmitWitnessTable(cls, traitName);
-  getOrCreateLlvmType(existentialType);
-  llvm::StructType* st = llvm::cast<llvm::StructType>(existentialType->getLlvmType());
-  llvm::Value* payload = builder->CreateBitCast(classPtrVal, builder->getPtrTy());
-  llvm::Value* wptr = builder->CreateBitCast(wit, builder->getPtrTy());
-  llvm::Value* u = llvm::UndefValue::get(st);
-  llvm::Value* s0 = builder->CreateInsertValue(u, payload, 0U);
-  return builder->CreateInsertValue(s0, wptr, 1U);
-}
-
-auto Codegen::callExistentialMethod(llvm::SMRange span, lesma::Value* receiver,
-                                    const std::string& methodName,
-                                    const std::vector<lesma::Value*>& args,
-                                    const std::vector<lesma::Type*>& explicitTypeArgs)
-    -> std::unique_ptr<lesma::Value> {
-  (void) explicitTypeArgs;
-  lesma::Type* receiverType = receiver->getType();
-  llvm::Value* fatVal = receiver->getLlvmValue();
-  if (receiverType->is(BaseType::TY_PTR) && receiverType->getElementType() != nullptr &&
-      receiverType->getElementType()->is(BaseType::TY_TRAIT_EXISTENTIAL)) {
-    fatVal = builder->CreateLoad(receiverType->getElementType()->getLlvmType(), fatVal);
-    receiverType = receiverType->getElementType();
-  }
-  if (!receiverType->is(BaseType::TY_TRAIT_EXISTENTIAL)) {
-    throw CodegenError(span, "Expected trait existential receiver for dynamic dispatch");
-  }
-  const std::string& traitName = receiverType->getDisplayName();
-  auto ordIt = traitRequirementMethodOrder.find(traitName);
-  if (ordIt == traitRequirementMethodOrder.end()) {
-    throw CodegenError(span, "Trait {} has no codegen metadata", traitName);
-  }
-  const auto& order = ordIt->second;
-  size_t idx = static_cast<size_t>(-1);
-  for (size_t i = 0; i < order.size(); ++i) {
-    if (order[i] == methodName) {
-      idx = i;
-      break;
-    }
-  }
-  if (idx == static_cast<size_t>(-1)) {
-    throw CodegenError(span, "Method {} not in trait {}", methodName, traitName);
-  }
-  const TraitDecl* tr = traitDeclByName[traitName];
-  const FuncDecl* req = findTraitRequirement(tr, methodName);
-  if (req == nullptr) {
-    throw CodegenError(span, "Trait {} missing requirement {}", traitName, methodName);
-  }
-
-  llvm::Value* payload = builder->CreateExtractValue(fatVal, 0U);
-  llvm::Value* tablePtr = builder->CreateExtractValue(fatVal, 1U);
-  llvm::Type* ptrTy = builder->getPtrTy();
-  llvm::ArrayType* arrTy = llvm::ArrayType::get(ptrTy, order.size());
-  llvm::Value* slot = builder->CreateInBoundsGEP(
-      arrTy, tablePtr, {builder->getInt32(0), builder->getInt32(static_cast<unsigned>(idx))});
-  llvm::Value* fnPtr = builder->CreateLoad(ptrTy, slot);
-
-  llvm::SmallVector<llvm::Type*, 8> tparams;
-  tparams.push_back(ptrTy);
-  for (Parameter* p : req->getParameters()) {
-    if (p->type != nullptr) {
-      p->type->accept(*this);
-      lesma::Type* pt = result->getType();
-      getOrCreateLlvmType(pt);
-      tparams.push_back(pt->getLlvmType());
-    }
-  }
-  req->getReturnType()->accept(*this);
-  lesma::Type* retLesma = result->getType();
-  getOrCreateLlvmType(retLesma);
-  llvm::Type* llvmRet = retLesma->getLlvmType();
-  llvm::FunctionType* callTy = llvm::FunctionType::get(llvmRet, tparams, false);
-  llvm::SmallVector<llvm::Value*, 8> callArgs;
-  callArgs.push_back(payload);
-  for (lesma::Value* a : args) {
-    callArgs.push_back(a->getLlvmValue());
-  }
-  if (llvmRet->isVoidTy()) {
-    builder->CreateCall(callTy, fnPtr, callArgs);
-    return std::make_unique<Value>(
-        "", cacheType(std::make_unique<Type>(BaseType::TY_VOID, builder->getVoidTy())), nullptr);
-  }
-  llvm::Value* ret = builder->CreateCall(callTy, fnPtr, callArgs);
-  return std::make_unique<Value>("", retLesma, ret);
-}
-
-auto Codegen::bindGenericsFromTypePair(const TypeExpr* declared, lesma::Type* actual,
-                                       const std::unordered_set<std::string>& genericNameSet,
-                                       std::unordered_map<std::string, lesma::Type*>& env) -> void {
-  if (declared == nullptr || actual == nullptr) {
-    return;
-  }
-  if (declared->getType() == TokenType::CUSTOM_TYPE) {
-    const std::string name = declared->getLookupName();
-    const auto declTypeArgs = declared->getTypeArgs();
-    lesma::Type* classActual = actual;
-    if (classActual->is(BaseType::TY_PTR) && classActual->getElementType() != nullptr &&
-        classActual->getElementType()->is(BaseType::TY_CLASS)) {
-      classActual = classActual->getElementType();
-    }
-    if (!declTypeArgs.empty() && classActual != nullptr && classActual->is(BaseType::TY_CLASS)) {
-      const std::string& display = classActual->getDisplayName();
-      std::string actualBase = display;
-      const auto anglePos = display.find('<');
-      if (anglePos != std::string::npos) {
-        actualBase = display.substr(0, anglePos);
-      }
-      if (name == actualBase) {
-        if (auto envIt = specializedClassTypeEnvs.find(classActual);
-            envIt != specializedClassTypeEnvs.end()) {
-          const Class* templateClass = nullptr;
-          if (auto git = genericClasses.find(name); git != genericClasses.end()) {
-            templateClass = git->second;
-          } else if (scope != nullptr) {
-            if (auto* sym = scope->lookupStruct(name);
-                sym != nullptr && sym->getGenericClassTemplate() != nullptr) {
-              templateClass = static_cast<const Class*>(sym->getGenericClassTemplate());
-            }
-          }
-          if (templateClass != nullptr) {
-            const std::vector<std::string> templateGenericParams =
-                templateClass->getGenericParams();
-            if (templateGenericParams.size() == declTypeArgs.size()) {
-              for (size_t i = 0; i < declTypeArgs.size(); ++i) {
-                if (auto concreteIt = envIt->second.find(templateGenericParams[i]);
-                    concreteIt != envIt->second.end() && concreteIt->second != nullptr) {
-                  bindGenericsFromTypePair(declTypeArgs[i], concreteIt->second, genericNameSet,
-                                           env);
-                }
-              }
-              return;
-            }
-          }
-        }
-      }
-    }
-    if (genericNameSet.contains(name) && !env.contains(name)) {
-      env[name] = actual;
-      return;
-    }
-    if (name == "__buffer" && actual->is(BaseType::TY_ARRAY) &&
-        actual->getElementType() != nullptr) {
-      if (declTypeArgs.size() == 1U) {
-        bindGenericsFromTypePair(declTypeArgs.front(), actual->getElementType(), genericNameSet,
-                                 env);
-      }
-    }
-    return;
-  }
-  if (declared->getType() == TokenType::PTR_TYPE && actual->is(BaseType::TY_PTR) &&
-      declared->getElementType() != nullptr && actual->getElementType() != nullptr) {
-    bindGenericsFromTypePair(declared->getElementType(), actual->getElementType(), genericNameSet,
-                             env);
-    return;
-  }
-  if (declared->getType() == TokenType::FUNC_TYPE && actual->is(BaseType::TY_FUNCTION)) {
-    if (declared->getReturnType() != nullptr && actual->getReturnType() != nullptr) {
-      bindGenericsFromTypePair(declared->getReturnType(), actual->getReturnType(), genericNameSet,
-                               env);
-    }
-    auto declParams = declared->getParams();
-    auto actualFields = actual->getFields();
-    for (size_t i = 0; i < declParams.size() && i < actualFields.size(); ++i) {
-      bindGenericsFromTypePair(declParams[i], actualFields[i]->type, genericNameSet, env);
-    }
-  }
-}
-
-auto Codegen::wrapNominalReturnAsPointer(Type* t) -> Type* {
-  if (t == nullptr) {
-    return nullptr;
-  }
-  if (t->is(BaseType::TY_PTR)) {
-    return t;
-  }
-  if (t->is(BaseType::TY_CLASS) || t->is(BaseType::TY_TRAIT_EXISTENTIAL)) {
-    return cacheType(std::make_unique<Type>(BaseType::TY_PTR, nullptr, t));
-  }
-  return t;
-}
-
-auto Codegen::specializeFunction(const FuncDecl* node, const std::vector<lesma::Type*>& paramTypes,
-                                 const std::vector<std::string>& genericNames,
-                                 const std::vector<lesma::Type*>& explicitTypeArgs)
-    -> lesma::Value* {
-  auto saved = currentGenericTypes;
-  std::unordered_map<std::string, lesma::Type*> env = saved;
-  if (!explicitTypeArgs.empty()) {
-    if (explicitTypeArgs.size() != genericNames.size()) {
-      throw CodegenError(
-          node->getSpan(),
-          "Explicit type argument count {} does not match generic parameter count {}",
-          explicitTypeArgs.size(), genericNames.size());
-    }
-    for (size_t i = 0; i < genericNames.size(); ++i) {
-      env[genericNames[i]] = explicitTypeArgs[i];
-    }
-  }
-  std::unordered_set<std::string> genericNameSet(genericNames.begin(), genericNames.end());
-  auto templateParams = node->getParameters();
-  size_t offset = (selfSymbol != nullptr) ? 1U : 0U;
-  for (size_t i = 0; i < templateParams.size() && (i + offset) < paramTypes.size(); ++i) {
-    TypeExpr* declType = templateParams[i]->type.get();
-    if (declType != nullptr) {
-      bindGenericsFromTypePair(declType, paramTypes[i + offset], genericNameSet, env);
-    }
-  }
-  currentGenericTypes = env;
-
-  for (auto* paramType : paramTypes) {
-    if (paramType != nullptr) {
-      getOrCreateLlvmType(paramType);
-    }
-  }
-  for (auto* explicitTypeArg : explicitTypeArgs) {
-    if (explicitTypeArg != nullptr) {
-      getOrCreateLlvmType(explicitTypeArg);
-    }
-  }
-  std::string key =
-      getMangledName(node->getSpan(), node->getName(), paramTypes, selfSymbol != nullptr);
-  if (auto it = specializedFunctions.find(key); it != specializedFunctions.end()) {
-    currentGenericTypes = std::move(saved);
-    return it->second;
-  }
-
-  std::vector<std::unique_ptr<Field>> fields;
-  std::vector<lesma::Type*> concreteParamTypes;
-  if (selfSymbol != nullptr) {
-    fields.push_back(std::make_unique<Field>("self", selfSymbol->getType()));
-    concreteParamTypes.push_back(selfSymbol->getType());
-  }
-  for (auto* param : node->getParameters()) {
-    param->type->accept(*this);
-    Type* paramT = result->getType();
-    if (paramT != nullptr && paramT->is(BaseType::TY_CLASS)) {
-      paramT = cacheType(std::make_unique<Type>(BaseType::TY_PTR, nullptr, paramT));
-    }
-    fields.push_back(std::make_unique<Field>(param->name, paramT));
-    concreteParamTypes.push_back(paramT);
-  }
-  node->getReturnType()->accept(*this);
-  Type* returnType = wrapNominalReturnAsPointer(result->getType());
-  std::vector<llvm::Type*> paramLLVMTypes;
-  for (auto* t : concreteParamTypes) {
-    getOrCreateLlvmType(t);
-    paramLLVMTypes.push_back(t->getLlvmType());
-  }
-  getOrCreateLlvmType(returnType);
-  auto funcType =
-      std::make_unique<Type>(BaseType::TY_FUNCTION, builder->getPtrTy(), std::move(fields));
-  funcType->setReturnType(returnType);
-  funcType->setVarArgs(node->getVarArgs());
-  auto* typePtr = cacheType(std::move(funcType));
-  auto mangledName =
-      getMangledName(node->getSpan(), node->getName(), concreteParamTypes, selfSymbol != nullptr);
-  auto func = std::make_unique<Value>(node->getName(), typePtr);
-  func->setCategory(ValueCategory::CALLABLE_SYMBOL);
-  func->setMangledName(mangledName);
-  func->setExported(node->isExported());
-  auto linkage = node->isExported() ? Function::ExternalLinkage : Function::PrivateLinkage;
-  llvm::Type* llvmReturnType = nullptr;
-  if (returnType->is(BaseType::TY_PTR) || returnType->is(BaseType::TY_CLASS)) {
-    llvmReturnType = builder->getPtrTy();
-  } else {
-    llvmReturnType = returnType->getLlvmType();
-  }
-  auto* llvmFuncType = FunctionType::get(llvmReturnType, paramLLVMTypes, node->getVarArgs());
-  auto* llvmFunc = Function::Create(llvmFuncType, linkage, mangledName, *theModule);
-  typePtr->setLlvmType(llvmFuncType);
-  func->setLlvmValue(llvmFunc);
-  auto* funcPtr = func.get();
-  scope->insertSymbol(std::move(func));
-  prototypes.emplace_back(funcPtr, node, selfSymbol);
-  specializedFunctions.emplace(std::move(key), funcPtr);
-  specializationEnvs.emplace(funcPtr, currentGenericTypes);
-  currentGenericTypes = std::move(saved);
-  return funcPtr;
-}
-
-auto Codegen::specializeClass(const Class* node,
-                              const std::vector<lesma::Type*>& constructorArgTypes,
-                              const std::vector<lesma::Type*>& explicitTypeArgs) -> lesma::Value* {
-  auto genericNames = node->getGenericParams();
-
-  const FuncDecl* constructorDecl = nullptr;
-  for (auto* method : node->getMethods()) {
-    if (method->getName() == "new") {
-      constructorDecl = method;
-      break;
-    }
-  }
-
-  std::unordered_map<std::string, lesma::Type*> env;
-  if (!explicitTypeArgs.empty()) {
-    if (explicitTypeArgs.size() != genericNames.size()) {
-      throw CodegenError(node->getSpan(),
-                         "Explicit type argument count {} does not match generic class parameter "
-                         "count {}",
-                         explicitTypeArgs.size(), genericNames.size());
-    }
-    for (size_t i = 0; i < genericNames.size(); ++i) {
-      env[genericNames[i]] = explicitTypeArgs[i];
-    }
-  }
-  std::unordered_set<std::string> genericNameSet(genericNames.begin(), genericNames.end());
-  if (constructorDecl != nullptr) {
-    auto params = constructorDecl->getParameters();
-    for (size_t i = 0; i < params.size() && i < constructorArgTypes.size(); ++i) {
-      TypeExpr* declType = params[i]->type.get();
-      if (declType != nullptr) {
-        bindGenericsFromTypePair(declType, constructorArgTypes[i], genericNameSet, env);
-      }
-    }
-  }
-
-  for (const auto& gn : genericNames) {
-    if (!env.contains(gn)) {
-      throw CodegenError(node->getSpan(),
-                         "Generic class {} requires type arguments for all parameters; "
-                         "could not infer {} from constructor",
-                         node->getIdentifier(), gn);
-    }
-  }
-
-  std::string key = (alias.empty() ? "" : "&" + alias + "=>") + node->getIdentifier();
-  for (const auto& gn : genericNames) {
-    key += "|" + MangleUtils::getTypeMangledName(node->getSpan(), env[gn]);
-  }
-  if (auto it = specializedClasses.find(key); it != specializedClasses.end()) {
-    return it->second;
-  }
-
-  auto saved = currentGenericTypes;
-  auto* savedSelfSymbol = selfSymbol;
-  currentGenericTypes = env;
-
-  for (auto* constructorArgType : constructorArgTypes) {
-    if (constructorArgType != nullptr) {
-      getOrCreateLlvmType(constructorArgType);
-    }
-  }
-  for (auto* explicitTypeArg : explicitTypeArgs) {
-    if (explicitTypeArg != nullptr) {
-      getOrCreateLlvmType(explicitTypeArg);
-    }
-  }
-
-  std::string concreteName = (alias.empty() ? "" : alias + "_") + node->getIdentifier();
-  for (const auto& gn : genericNames) {
-    concreteName += "_" + MangleUtils::getTypeMangledName(node->getSpan(), env[gn]);
-  }
-
-  std::vector<std::unique_ptr<Field>> fields;
-  std::vector<llvm::Type*> elementLLVMTypes;
-  for (auto* field : node->getFields()) {
-    if (field->getType() != nullptr) {
-      field->getType()->accept(*this);
-    } else {
-      field->getValue()->accept(*this);
-    }
-    getOrCreateLlvmType(result->getType());
-    elementLLVMTypes.push_back(result->getType()->getLlvmType());
-    std::unique_ptr<Value> defaultVal;
-    if (field->getValue() != nullptr) {
-      defaultVal = std::move(result);
-      if (field->getType() != nullptr) {
-        field->getType()->accept(*this);
-      } else {
-        result = std::make_unique<Value>(*defaultVal);
-      }
-    }
-    fields.push_back(std::make_unique<Field>(field->getIdentifier()->getValue(), result->getType(),
-                                             std::move(defaultVal)));
-  }
-
-  auto* structType =
-      llvm::StructType::create(theModule->getContext(), elementLLVMTypes, concreteName);
-  auto type = std::make_unique<Type>(BaseType::TY_CLASS, structType, std::move(fields));
-  std::string displayName = node->getIdentifier() + "<";
-  for (size_t i = 0; i < genericNames.size(); ++i) {
-    if (i > 0U) {
-      displayName += ", ";
-    }
-    displayName += env[genericNames[i]]->toString();
-  }
-  displayName += ">";
-  type->setDisplayName(displayName);
-  type->setImplTraitNames(std::vector<std::string>(node->getImplTraitNames()));
-  auto* typePtr = type.get();
-  scope->insertType(concreteName, std::move(type));
-
-  auto structSymbol = std::make_unique<Value>(concreteName, typePtr);
-  structSymbol->setCategory(ValueCategory::TYPE_SYMBOL);
-  structSymbol->setExported(node->isExported());
-  auto* structSymbolPtr = structSymbol.get();
-  scope->insertSymbol(std::move(structSymbol));
-  specializedClasses.emplace(key, structSymbolPtr);
-  specializedClassSymbolsByType[typePtr] = structSymbolPtr;
-  specializedClassTypeEnvs[typePtr] = env;
-
-  auto* selfType =
-      cacheType(std::make_unique<Type>(BaseType::TY_PTR, builder->getPtrTy(), typePtr));
-  methodSelfSymbols.push_back(std::make_unique<Value>(concreteName, selfType));
-  methodSelfSymbols.back()->setExported(node->isExported());
-  selfSymbol = methodSelfSymbols.back().get();
-
-  auto hasConstructor = false;
-  for (auto* method : node->getMethods()) {
-    method->accept(*this);
-    if (method->getName() == "new") {
-      hasConstructor = true;
-      std::vector<lesma::Type*> constructorParams = {selfSymbol->getType()};
-      auto* constructor = scope->lookupFunction("new", constructorParams);
-      structSymbolPtr->setConstructor(constructor);
-    }
-  }
-  selfSymbol = nullptr;
-
-  if (!hasConstructor) {
-    throw CodegenError(node->getSpan(), "Generic class {} has no constructors",
-                       node->getIdentifier());
-  }
-  selfSymbol = savedSelfSymbol;
-  currentGenericTypes = std::move(saved);
-  return structSymbolPtr;
-}
 
 auto Codegen::visit(const Compound* node) -> void {
   for (auto* elem : node->getChildren()) {
@@ -3950,7 +1924,7 @@ auto Codegen::visit(const IsOp* node) -> void {
   bool typesEqual = leftType->isEqual(rightType);
   if (!typesEqual && leftType != nullptr && rightType != nullptr &&
       leftType->getBaseType() == rightType->getBaseType() &&
-      leftType->isOneOf({BaseType::TY_CLASS, BaseType::TY_ARRAY, BaseType::TY_ENUM})) {
+      TypeUtils::isNominalTypeForIdentity(leftType)) {
     typesEqual = leftType->toString() == rightType->toString();
   }
 
@@ -4030,7 +2004,8 @@ auto Codegen::visit(const ListLiteral* node) -> void {
     if (fields.empty() || fields.front()->type == nullptr ||
         !fields.front()->type->is(BaseType::TY_ARRAY) ||
         fields.front()->type->getElementType() == nullptr) {
-      throw CodegenError(node->getSpan(), "List literal resolved to invalid stdlib list class");
+      throw CodegenError(node->getSpan(),
+                         "List literal resolved to invalid class (expected __buffer<T> field)");
     }
     auto* bufferType = fields.front()->type;
     auto* structType = cast<llvm::StructType>(getOrCreateLlvmType(listType));
@@ -4537,6 +2512,13 @@ auto Codegen::callNamedFunction(llvm::SMRange span, const std::string& functionN
   selfSymbol = selfSymbolTmp;
   return std::make_unique<Value>("", symbol->getType()->getReturnType(),
                                  builder->CreateCall(func, localParamsLLVM));
+}
+
+auto Codegen::isBuiltinListBuiltinMethodName(const std::string& methodName) const -> bool {
+  return methodName == "len" || methodName == "clear" || methodName == "push" ||
+         methodName == "pop" || methodName == "copy" ||
+         methodName == std::string{OperatorUtils::SUBSCRIPT_GET_NAME} ||
+         methodName == std::string{OperatorUtils::SUBSCRIPT_SET_NAME};
 }
 
 auto Codegen::callListMethodByName(llvm::SMRange span, lesma::Value* receiver,
