+++ src/liblesma/Symbol/TypeUtils.cpp
@@ -29,4 +29,19 @@ auto findFieldInFields(Type* structType, const std::string& field) -> Field* {
   }
   return nullptr;
 }
+
+auto isNominalTypeForIdentity(Type const* t) -> bool {
+  return t != nullptr &&
+         t->isOneOf({BaseType::TY_CLASS, BaseType::TY_ARRAY, BaseType::TY_ENUM});
+}
+
+auto passesByPointerInAbi(Type const* t) -> bool {
+  if (t == nullptr) {
+    return false;
+  }
+  if (t->is(BaseType::TY_PTR)) {
+    return true;
+  }
+  return t->is(BaseType::TY_CLASS) || t->is(BaseType::TY_TRAIT_EXISTENTIAL);
+}
 } // namespace lesma::TypeUtils
